GraphFigExamples

Description

Documentation

padBorder :: Diagram B -> Diagram B

used for most diagrams to give border padding

Example Tgraphs with Figures

fool :: Tgraph

fool: fool's kite - a decomposed left and right kite back-to-back (i.e. not sharing join edge)

foolD :: Tgraph

a once decomposed fool (= foolDs!!1)

foolDminus :: Tgraph

foolDminus: 3 faces removed from foolD

foolDs :: [Tgraph]

a list of all decompositions of fool

foolFig :: Diagram B

diagram of just fool

foolAndFoolD :: Diagram B

diagram of fool with foolD

sunGraph :: Tgraph

Tgraph for a sun

sunDs :: [Tgraph]

All decompositions of sunGraph

figSunD3D2 :: Diagram B

Figure for a 3 times decomposed sun with a 2 times decomposed sun

figSunD2D :: Diagram B

Fifure for a 2 times decomposed sun with a once decomposed sun

kiteGraph :: Tgraph

Tgraph for kite

kiteDs :: [Tgraph]

All decompositions of a kite

dartGraph :: Tgraph

Tgraph for a dart

dartDs :: [Tgraph]

All decompositions of a dart

Partial Compositions figures

checkPCompose :: Tgraph -> Diagram B

Shows labelled vertices For checking partCompose to get alignment vertices

showPCompose :: Tgraph -> (Vertex, Vertex) -> Diagram B

showPCompose g (a,b) applies partCompose to g, then aligns and draws the composed graph with the remainder faces (in lime) It can cope with an empty composed graph. Vertices a and b must be common to composed g and g (use checkPCompose to choose vertices)

pCompFig1 :: Diagram B

diagrams showing partial compositions (with ignored faces in pale green)

pCompFig2 :: Diagram B

diagrams showing partial compositions (with ignored faces in pale green)

pCompFig :: Diagram B

diagrams showing partial compositions (with ignored faces in pale green)

Forced Tgraph figures

forceFoolDminus :: Diagram B

diagram of foolDminus and the reult of forcing

drawForce :: Tgraph -> Diagram B

new version of drawForce using subTgraphs instead of alignments

forceDartD3Fig :: Diagram B

diagrams of forced graphs (3 or 5 times decomposed kite or dart or sun)

forceDartD5Fig :: Diagram B

diagrams of forced graphs (3 or 5 times decomposed kite or dart or sun)

forceKiteD3Fig :: Diagram B

diagrams of forced graphs (3 or 5 times decomposed kite or dart or sun)

forceKiteD5Fig :: Diagram B

diagrams of forced graphs (3 or 5 times decomposed kite or dart or sun)

forceSunD5Fig :: Diagram B

diagrams of forced graphs (3 or 5 times decomposed kite or dart or sun)

forceFig :: Diagram B

diagrams of forced graphs (3 or 5 times decomposed kite or dart or sun)

Emplacements with figures

drawEmbed :: Tgraph -> Tgraph -> (Vertex, Vertex) -> Diagram B

drawEmbed (a,b) g1 g2 embeds g1 (coloured red) onto g2, by aligning with (a,b) vertices a and b must be common to g1 and g2

drawEmbedVP :: VPatch -> VPatch -> (Vertex, Vertex) -> Diagram B

drawEmbedVP (a,b) vp1 vp2 embeds vp1 (coloured red) onto vp2, after aligning with (a,b) vertices a and b must be common to vp1 and vp2

checkEmbed :: Tgraph -> Tgraph -> Diagram B

use checkEmbed to find a common pair of vertices to align (for use with drawEmbed)

lowestJoin :: Tgraph -> (Vertex, Vertex)

returns the join edge with lowest origin (and lowest oppV of faces with that origin)

drawMaxEmplace :: Tgraph -> Diagram B

drawMaxEmplace g produces the emplacement of g overlaid with the maximal (forced) composition of g

maxEmplaceFig :: Diagram B

an example showing the maximum composition (a kite) over 4 times decomposed pair of darts

Multi emplace and choices

foolChoices :: Diagram B

four choices for composing fool

emplaceFoolDChoices :: Diagram B

emplacement choices for foolD WARNING: relies on numbering which can change with changes to forcing Vertex 1 is not present in the final choice (hence 29)

Removed faces (forcing and composing)

dartD4 :: Tgraph

Tgraph of 4 times decomposed dartGraph

brokenDart :: Tgraph

brokenDart gets repaired by forcing but can also be composed to a maximal graph

badlyBrokenDart :: Tgraph

badlyBrokenDart gets repaired by forcing but will fail to produce a valid graph if composed twice without forcing *** Exception: checkTgraph: crossing boundaries found at [3] in Tgraph {vertices = [4,6,3,1,5], faces = [LD (4,6,3),LK (1,5,3)]}

brokenDartFig :: Diagram B

brokenDartFig shows the faces removed from dartD4 to make brokenDart and badlyBrokenDart

checkBrokenDartFig :: Diagram B

the force of the worst case (badlyBrokenDart)

brokenKitesDFig :: Diagram B

Showing the decomposed pair of adjacent kites, followed by brokenKites (3 faces removed from decomposed pair of kites), followed by compositon of brokenKites (a kite) which is also composition of forced brokenkites, followed by the emplacement of brokenKites The figure illustrates that brokenKites is included in emplace brokenKites even though the missing part is not repaired.

brokenKites :: Tgraph

2 adjacent kites decomposed then the top half kite components removed (3 of them)

touchingTestFig :: Diagram B

diagram of touching vertex situation and forced result

Incorrect Tgraphs (and other problem Tgraphs)

crossingBdryFig :: Diagram B

faces removed from foolD to illustrate crossing boundary and non-face-connected VPatches

mistake :: Tgraph

mistake is an erroneous graph with a kite bordered by 2 darts

pfMistakeFig :: Diagram B

mistake and the point at which forcing fails

forcingDmistakeFig :: Diagram B

decomposeG mistake and the point at which forcing fails with RK (6,26,1)

forcingD2mistakeFig :: Diagram B

forcingD2mistakeFig Figure showing a stuck graph with error at vertex 35 This is the result of twice decomposed mistake which fails when forced *** Exception: errorcheckClock: Found incorrect graph Conflict at vertex: 35 forcingD2mistakeFig shows the graph when the error is discovered.

mistake1 :: Tgraph

mistake1 is a kite bordered by 2 half kites (subgraph of mistake and still erroneous)

partFMistake1Fig :: Diagram B

partially forced mistake1 (at the point of discovery of incorrect graph

cdMistake1Fig :: Diagram B

decomposed mistake1 is no longer erroneous and can be forced and recomposed

Figures fof 7 vertex types

vertexTypesFig :: QDiagram B V2 Double Any

vertexTypesFig is 7 vertex types single diagram as a row

labelD :: String -> Diagram B -> Diagram B

add a given label below and to right of centre of a diagram

jackGraph :: Tgraph

Graphs for the vertex types (other than sunGraph previously declared) 7 vertex types are, sunGraph, starGraph, jackGraph, queenGraph, kingGraph, aceGraph=fool, deuceGraph

kingGraph :: Tgraph

Graphs for the vertex types (other than sunGraph previously declared) 7 vertex types are, sunGraph, starGraph, jackGraph, queenGraph, kingGraph, aceGraph=fool, deuceGraph

queenGraph :: Tgraph

Graphs for the vertex types (other than sunGraph previously declared) 7 vertex types are, sunGraph, starGraph, jackGraph, queenGraph, kingGraph, aceGraph=fool, deuceGraph

aceGraph :: Tgraph

Graphs for the vertex types (other than sunGraph previously declared) 7 vertex types are, sunGraph, starGraph, jackGraph, queenGraph, kingGraph, aceGraph=fool, deuceGraph

deuceGraph :: Tgraph

Graphs for the vertex types (other than sunGraph previously declared) 7 vertex types are, sunGraph, starGraph, jackGraph, queenGraph, kingGraph, aceGraph=fool, deuceGraph

starGraph :: Tgraph

Graphs for the vertex types (other than sunGraph previously declared) 7 vertex types are, sunGraph, starGraph, jackGraph, queenGraph, kingGraph, aceGraph=fool, deuceGraph

forceVFigures :: [Diagram B]

forceVFigures is a list of 7 diagrams - force of 7 vertex types

forceVsFig :: Diagram B

forceVsFig shows emplacements of 7 vertex types in a row as single diagram

relatedVTypeFig :: QDiagram B V2 Double Any

relatedVTypeFig lays out figures from forceVFigures plus a kite as single diagram with 3 columns

Other miscelaneous Tgraphs and Diagrams

boundaryFDart4 :: Tgraph

graphs of the boundary faces only of forced graphs (dartDs!!4 and dartDs!!5)

boundaryFDart5 :: Tgraph

graphs of the boundary faces only of forced graphs (dartDs!!4 and dartDs!!5)

boundaryGapFDart4 :: Tgraph

graphs of the boundary faces only of a forced graph - with extra faces removed to make a gap

boundaryGapFDart5 :: Tgraph

graphs of the boundary faces only of a forced graph - with extra faces removed to make a gap

boundaryGap4Fig :: Diagram B

figures for the boundary gap graphs boundaryGapFDart4, boundaryGapFDart5

boundaryGap5Fig :: Diagram B

figures for the boundary gap graphs boundaryGapFDart4, boundaryGapFDart5

gapProgress5 :: Diagram B

showing intermediate state of filling the inlet and closing the gap of boundaryGapFDart5 using stepForce 2000

gapProgress4 :: Diagram B

showing intermediate state of filling the inlet and closing the gap of boundaryGapFDart4 using stepForce 600

bigPic0 :: Diagram B

bigPic is a diagram illustrating force/emplacement relationships for decomposed darts bigPic0 is main diagram for bigPic without the arrows

bigPic :: Diagram B

bigPic is a diagram illustrating force/emplacement relationships for decomposed darts bigPic0 is main diagram for bigPic without the arrows

forceArrow :: (IsName n1, IsName n2) => n1 -> n2 -> Diagram B -> Diagram B

add a force arrow (black) from named parts of 2 diagrams

forceDecArrow :: (IsName n1, IsName n2) => n1 -> n2 -> Diagram B -> Diagram B

add a (force . decompose) arrow (blue) from named parts of 2 diagrams

decompArrow :: (IsName n1, IsName n2) => n1 -> n2 -> Diagram B -> Diagram B

add a decompose arrow (ddashed blue) from named parts of 2 diagrams

composeArrow :: (IsName n1, IsName n2) => n1 -> n2 -> Diagram B -> Diagram B

add a compose arrow (green) from named parts of 2 diagrams

composeArc :: (IsName n1, IsName n2) => n1 -> n2 -> Diagram B -> Diagram B

add a compose arrow arc (green) from named parts of 2 diagrams

curioPic0 :: Diagram B

curioPic0 is diagram curioPic without arrows

curioPic :: Diagram B

curioPic is a diagram illustrating where compose looses information not recovered by force with sunPlus3Dart' third item in bottom row, (curioPic0 is diagram without arrows)

graphOrder1 :: QDiagram B V2 Double Any

figure showing ordering of a decomposed kite (bottom), a test graph with an extra LK(3,6,8), and forced figure at the top and composition of all 3 = kite on the right

Testing (functions and figures and experiments)

testForce4 :: Diagram B

diagrams of forced graphs for boundaryGapFDart4 and boundaryGapFDart5

testForce5 :: Diagram B

diagrams of forced graphs for boundaryGapFDart4 and boundaryGapFDart5

testViewBoundary :: [Int] -> Boundary -> Diagram B

testViewBoundary is a testing tool to inspect the boundary vertex locations of some (intermediate) Boundary -- (used in conjunction with stepForce to get an intermediate Boundary) -- The boundary edges of a Boundary are shown in lime - using the Boundary positions of vertices -- The graph is converted to a vp separately (so using a fresh calculation of positions) -- Thus rotations may be needed to match up. -- Use an empty list of integer rotations to see what rotations are needed to align the figures.

inspectForce5 :: Int -> Diagram B

used to discover accuracy problem of older thirdVertexLoc view tha boundary after n steps of forcing (starting with boundaryGapFDart5) e.g. n = 1900

inspectForce3 :: Int -> Diagram B

used to discover accuracy problem of older thirdVertexLoc view tha boundary after n steps of forcing (starting with boundaryGapFDart5) e.g. n = 1900

testBoundary4 :: Diagram B

figures showing boundary edges of the boundary gap graphs boundaryGapFDart4 and boundaryGapFDart5

testBoundary5 :: Diagram B

figures showing boundary edges of the boundary gap graphs boundaryGapFDart4 and boundaryGapFDart5

testCrossingBoundary :: Tgraph

testing crossing boundary detection e.g. by using force on testCrossingBoundary

checkCompleteFig :: Diagram B

test wholeTiles (which adds missing second halves of each face)

checkGraphFromVP :: Diagram B

test graphFromVP

dartsOnlyFig :: Diagram B

figure testing selectFacesGtoVP by removing all kites

Testing SubTgraphs

subExample :: SubTgraph

subExample is a SubTgraph of a forced 5 times decomposed dart with tracked faces from a forced 2 times decomposed dart

subExampleFig :: Diagram B

subExampleFig draws subExample with the tracked faces in red

hollowGraph :: Tgraph

hollowGraph happens to be a valid Tgraph after removing the tracked faces from subExample This is not generally the case

forceHollowFig :: Diagram B

figure showing hollowGraph and result of forcing

drawWithoutTracked :: SubTgraph -> Diagram B

drawing non tracked faces in general

removeTrackedFig :: Diagram B

example using drawWithoutTracked

Viewing choice results with SubTgraphs

checkChoiceEdge :: Tgraph -> QDiagram B V2 Double Any

for Tgraph g produce diagram of forced version showing vertices (to inspect edges for selection)

trackTwoChoices :: Tgraph -> DEdge -> [SubTgraph]

given a boundary directed edge of a forced graph (either direction) construct two SubTgraphs with half dart/kite addition on the edge respectively track the resulting faces and also the singleton new face, then force both SubTgraphs

twoChoices :: [SubTgraph]

Take a forced, 4 times decomposed dart, then track the two choices

twoChoicesFig :: Diagram B

show the (tracked) twoChoices with drawSubTgraph2 (tracked faces in red, new face filled black)

moreChoices0 :: [SubTgraph]

track two further choices with the first of twoChoices (fullgraph)

moreChoices1 :: [SubTgraph]

track two further choices with the second of twoChoices (fullgraph)

moreChoicesFig0 :: Diagram B

figures for 4 further choices

moreChoicesFig1 :: Diagram B

figures for 4 further choices

moreChoicesFig :: Diagram B

figures for 4 further choices

dartHalfDart :: Tgraph

a dart with another half dart on a long edge

Trying to find which extensions to the starting dart correspond to the twoChoicesFig

dartHalfKite :: Tgraph

a dart with a half kite on a long edge

Trying to find which extensions to the starting dart correspond to the twoChoicesFig

dartPlusDart :: Tgraph

two darts sharing a long edge

Trying to find which extensions to the starting dart correspond to the twoChoicesFig

dartPlusKite :: Tgraph

a dart and a kite sharing a long edge

Trying to find which extensions to the starting dart correspond to the twoChoicesFig

kitePlusKite :: Tgraph

two kites sharing a long edge

sunPlusDart :: Tgraph

A sun with a single complete dart on the boundary

sunPlus2Dart :: Tgraph

A sun with 2 darts adjacent on the boundary

sunPlus2Dart' :: Tgraph

A sun with 2 darts NOT adjacent on the boundary

sunPlus3Dart :: Tgraph

A sun with 3 darts adjacent on the boundary

sunPlus3Dart' :: Tgraph

A sun with 3 darts on the boundary NOT all adjacent This example has an emplacement that does not include the original but is still a correct Tgraph

halfWholeFig :: Diagram B

halfWholeFig shows that a whole dart/kite needs to be added to get the same result as twoChoicesFig Adding a half tile has no effect on the forced decomposition

kkEmpsFig :: Diagram B

two kites (force decomp twice) figure

maxShapesFig :: Diagram B

two kites added to related vertex types figure

drawForceEmplace :: Tgraph -> QDiagram B V2 Double Any

drawForceEmplace g is a diagram for g followed by force g followed by emplace g

testD' :: QDiagram B V2 Double Any

force after adding half dart (rocket tip) to sunPlus3Dart'. Adding a kite half gives an incorrect graph discovered by forcing.

emplaceProblemFig :: QDiagram B V2 Double Any

sunPlus3Dart' is a sun with 3 darts on the boundary NOT all adjacent This example has an emplacement that does not include the original but is still a correct Tgraph. The figure shows the force and emplace difference.