HalfTile

Description

Documentation

data HalfTile rep

Representing Half Tile Pieces Polymorphicly Common code for both graphs and vector representations of tiling For vectors - rep is V2 Double For Tgraphs rep is (Vertex,Vertex,Vertex) (HalfTile is a functor)

Constructors

LD rep
Left Dart

RD rep
Right Dart

LK rep
Left Kite

RK rep
Right Kite

tileRep :: HalfTile rep -> rep

return the representation of a half-tile

isLD :: HalfTile rep -> Bool

half-tile predicates

isRD :: HalfTile rep -> Bool

half-tile predicates

isLK :: HalfTile rep -> Bool

half-tile predicates

isRK :: HalfTile rep -> Bool

half-tile predicates

isDart :: HalfTile rep -> Bool

half-tile predicates

isKite :: HalfTile rep -> Bool

half-tile predicates