Tgraph.Compose

Description

This module includes the main composition operation composeG but also exposes classifyDartWings (and type DWClass) for debugging and experimenting

Documentation

composeG :: Tgraph -> Tgraph

The main deterministic function for composing is composeG which is essentially partCompose after uncomposed faces are ignored. If the result fails to be connected or has crossing boundaries an error is raised.

partCompose :: Tgraph -> ([TileFace], Tgraph)

partCompose produces a Tgraph by composing faces which uniquely compose, returning a pair consisting of unused faces of the original graph along with the composed Tgraph. It checks the Tgraph for connectedness and no crossing boundaries raising an error if this check fails. (it makes use of classifyDartWings which also returns an association of faces incident with each dart wing so these do not need to be recalculated.)

data DWClass

DWClass is a record type for the result of classifying dart wings in a Tgraph. It also records vGroup - a mapping of each dart wing vertex in the Tgraph to the faces incident with the vertex. This is to save recalculating the face groups when composing with partCompose.

Constructors

DWClass

Fields

largeKiteCentres :: [Vertex]

largeDartBases :: [Vertex]

unknowns :: [Vertex]

vGroup :: Map Vertex [TileFace]

classifyDartWings :: Tgraph -> DWClass

classifyDartWings classifies all dart wing tips. The result is a DWClass record of largeKiteCentres, largeDartBases, unknowns and a mapping where largeKiteCentres are new kite centres, largeDartBases are new dart bases unknowns cannot be classified, and the mapping identifies the faces found at each vertex in both largeKiteCentres and largeDartBases. This information is recorded to make partCompose more efficient.