Tgraph.Convert

Description

Includes conversion operations from Tgraphs to diagrams as well as the intermediate type VPatch to allow vertex labels to be drawn.

VPatches

data DualRep

a DualRep contains two representations - a vector and a face(= 3 vertices)

Constructors

DualRep

Fields

vector :: V2 Double

face :: (Vertex, Vertex, Vertex)

dualRep :: V2 Double -> (Vertex, Vertex, Vertex) -> DualRep

construct a dualRep from a vector and a vertex triple

type Hybrid = HalfTile DualRep

A hybrid is a HalfTile with the dual representation of the face vertices and the join vector

data VPatch

A VPatch (vertex patch) is a patch with both face and vertex and vector information. It contains a list of Located vertices and a list of Located Hybrids

Constructors

VPatch

Fields

lVertices :: [Located Vertex]

lHybrids :: [Located Hybrid]

makeVPatch :: Tgraph -> VPatch

For converting a Tgraph to a VPatch. An empty graph is a special case. Otherwise it uses chooseLowest to choose a starting face (placed at the front of the list of faces) Then uses Tgraph.Prelude.createVPoints to form a mapping of vertices to positions. This makes the join of the face with lowest origin and lowest oppV align on the positive x axis

chooseLowest :: [TileFace] -> [TileFace]

For a non-empty list of tile faces find the face with lowest originV (and then lowest oppV). Move this face to the front of the returned list of faces. Used by makeVPatch (and hence makePatch) for non-empty Tgraphs

makePatch :: Tgraph -> Patch

makePatch uses makeVPatch first then the Hybrids are converted to Pieces and the Located Vertex information is dropped

graphFromVP :: VPatch -> Tgraph

An inverse to makeVPatch which checks for connected and no crossing boundaries

Drawing VPatches and Graphs

drawGraph :: Tgraph -> Diagram B

simplest drawing without vertex labels

dashJGraph :: Tgraph -> Diagram B

simplest drawing without vertex labels

drawVGraph :: Tgraph -> Diagram B

simplest drawing with vertex labels

drawVPatch :: VPatch -> Diagram B

convert a VPatch to a diagram with vertex labels

drawVPatchWith :: (Piece -> Diagram B) -> VPatch -> Diagram B

drawVPatchWith pd vp - convert VPatch vp to a diagram with vertex labels using pd to draw pieces

drawVlabels :: [Located Vertex] -> Diagram B

make a diagram of vertex labels given located vertices (used by drawVPatch and drawVPatchWith)

removeFacesVP :: [TileFace] -> VPatch -> VPatch

remove a list of faces from a VPatch

selectFacesVP :: [TileFace] -> VPatch -> VPatch

make a new VPatch with a list of selected faces from a VPatch

selectFacesGtoVP :: [TileFace] -> Tgraph -> VPatch

selectFacesGtoVP fcs g - only selected faces (fcs) are kept after converting g to a VPatch

removeFacesGtoVP :: [TileFace] -> Tgraph -> VPatch

removeFacesGtoVP fcs g - remove faces (fcs) after converting g to a VPatch

Alignment with Vertices

centerOn :: Vertex -> VPatch -> VPatch

center a VPatch on a particular vertex

alignXaxis :: (Vertex, Vertex) -> VPatch -> VPatch

alignXaxis takes a vertex pair (a,b) and a VPatch vp for centering vp on a and rotating the result so that b is on the positive X axis.

alignments :: [(Vertex, Vertex)] -> [VPatch] -> [VPatch]

alignments takes a list of vertex pairs for respective rotations of VPatches in the second list. For a pair (a,b) the Vpatch is centered on a then b is aligned along the positive x axis. The vertex pair list can be shorter than the list of vpatches - the remaining vpatches are left as they are.

alignAll :: (Vertex, Vertex) -> [VPatch] -> [VPatch]

alignAll (a,b) vpList provided both vertices a and b exist in each Vpatch in vpList, the VPatches are all aligned centred on a, with b on the positive x axis

Rotating and Scaling lists

rotations :: (Transformable a, V a ~ V2, N a ~ Double) => [Int] -> [a] -> [a]

rotations takes a list of integers (ttangles) for respective rotations of items in the second list (things to be rotated). This includes Diagrams, Patches, VPatches The integer list can be shorter than the list of items - the remaining items are left unrotated.

scales :: (Transformable a, V a ~ V2, N a ~ Double) => [Double] -> [a] -> [a]

scales takes a list of doubles for respective scalings of items in the second list (things to be scaled). This includes Diagrams, Patches, VPatches The list of doubles can be shorter than the list of items - the remaining items are left unscaled.

phiScales :: (Transformable a, V a ~ V2, N a ~ Double) => [a] -> [a]

increasing scales by phi along a list starting with 1

phiScaling :: (Transformable a, V a ~ V2, N a ~ Double) => Double -> [a] -> [a]

increasing scales by phi along a list starting with given first argument

Auxiliary definitions

vertexLocs :: VPatch -> [(Vertex, Point V2 Double)]

makes an association list of vertex to location from a VPatch

findLoc :: Vertex -> VPatch -> Maybe (Point V2 Double)

find the location of a specific vertex in a VPatch

withHybs :: ([Located Hybrid] -> [Located Hybrid]) -> VPatch -> VPatch

Apply a function to just the list of located hybrids in a VPatch (leaving located vertices untouched)

asPiece :: Hybrid -> Piece

convert a Hybrid to a Piece, dropping the Vertex information

asFace :: Hybrid -> TileFace

convert a Hybrid to a TileFace, dropping the Vector information

dropVertices :: VPatch -> Patch

dropVertices removes vertex information from Hybrids and removes located vertex list

dropVectors :: VPatch -> [TileFace]

dropVectors removes vector information from Hybrids and removes located vertex list

drawGBoundary :: Tgraph -> Diagram B

displaying the boundary of a Tgraph in lime

drawEdges :: Mapping Vertex (Point V2 Double) -> [(Vertex, Vertex)] -> Diagram B

produce a diagram of a list of edges (given a mapping of vertices to locations)

drawEdge :: Mapping Vertex (Point V2 Double) -> (Vertex, Vertex) -> Diagram B

produce a diagram of a single edge (given a mapping of vertices to locations)

Drawing SubTgraphs

drawSubTgraph :: [Patch -> Diagram B] -> SubTgraph -> Diagram B

To draw a SubTgraph, we need a list of functions turning patches into diagrams The first function is applied to a patch for untracked faces Subsequent functions are applied to the respective tracked subsets (Each patch is atop earlier ones, so the untracked patch is at the bottom)

drawSubTgraph1 :: SubTgraph -> Diagram B

special case of drawSubTgraph using 1 subset (and 2 patchdrawing functions): normal (black), then red

drawSubTgraph2 :: SubTgraph -> Diagram B

special case of drawSubTgraph using 2 subsets (and 3 patchdrawing functions): normal (black), then red, then filled black

Touching Vertex global check

touchingVertices :: Tgraph -> [(Vertex, Vertex)]

A TESTING function for use if the touching vertex check is switched off in forcing. This is a reptrospective check. touchingVertices checks that no vertices are too close to each other by making a VPatch of a Tgraph. If vertices are too close that indicates we may have the same point with two different vertex numbers arising from the touching vertex problem. It returns pairs of vertices that are too close (i.e less than 0.5 where 1.0 would be the length of short edges) An empty list is returned if there is no touching vertex problem. Complexity has order of the square of the number of vertices (calculates distance between all pairs)