Tgraph.Decompose

Description

This module includes the main decomposition operation decomposeG but also exposes some auxilliary functions for debugging, experimenting and use with SubTgraphs

Documentation

decomposeG :: Tgraph -> Tgraph

Decompose a Tgraph. This is uniquely determined.

newVPhiMap :: Tgraph -> ([Vertex], (Vertex, Vertex) -> Vertex)

newVPhiMap g produces newVs - a list of new vertices (one for each phi edge of v) and a function mapping each phi edge to its assigned vertex in newVs. Both (a,b) and (b,a) get the same v

decompFace :: ((Vertex, Vertex) -> Vertex) -> TileFace -> [TileFace]

Decompose a face producing new faces. This requires a function to get the unique vertex assigned to each phi edge (as created by newVPhiMap)

decompositionsG :: Tgraph -> [Tgraph]

infinite list of decompositions of a Tgraph

decomposeSub :: SubTgraph -> SubTgraph

decompose a SubTgraph - applies decomposition to all tracked subsets as well as the full Tgraph