Tgraph.Force

Description

This module includes force plus related operations for testing and experimenting. It also exposes the calculation of relative angle of edges at vertices used to find existing edges.

Touching vertex checking

touchCheckOn :: Bool

touchCheckOn is a global variable used to switch on/off checks for touching vertices when new vertices are added to a graph. Used by touchCheck. This should be True for safety. If it is switched to False, all results of forcing need to be checked for touching vertices.

touchCheck :: Point V2 Double -> Mapping a (Point V2 Double) -> Bool

check if a vertex location touches (is too close to) any other vertex location in the mapping

tooClose :: Point V2 Double -> Point V2 Double -> Bool

check if two vertex locatioons are too close (indicating touching vertices)

Boundary operations

data Boundary

A Boundary records the boundary directed edges (directed so that faces are on LHS and exterior is on RHS) plus a mapping of boundary vertices to their incident faces, plus a mapping of boundary vertices to positions. It also keeps track of all the faces and vertices, and the next vertex label to be used when adding a new vertex. Note that bvFacesMap is initially only defined for boundary vertices, but the information is not removed when a vertex is no longer on the boundary (after an update). Similarly for bvLocMap.

Constructors

Boundary

Fields

bDedges :: [DEdge]

boundary directed edges (face on LHS, exterior on RHS)

bvFacesMap :: Mapping Vertex [TileFace]

faces at each boundary vertex

bvLocMap :: Mapping Vertex (Point V2 Double)

position of each boundary vertex

allFaces :: [TileFace]

all the tile faces

allVertices :: [Vertex]

all the vertices

nextVertex :: Vertex

next vertex number

makeBoundary :: Tgraph -> Boundary

Calculates Boundary information from a Tgraph

recoverGraph :: Boundary -> Tgraph

Converts a Boundary back to a Tgraph

changeVFMap :: TileFace -> Mapping Vertex [TileFace] -> [Vertex] -> Mapping Vertex [TileFace]

changeVFMap f vfmap vs - adds f to the list of faces associated with each v in vs and updates vfmap

facesAtBV :: Boundary -> Vertex -> [TileFace]

facesAtBV bd v - returns the faces found at v (which must be a boundary vertex)

boundaryFaces :: Boundary -> [TileFace]

return the (set of) faces which have a boundary vertex from boundary information

Updates and ForceState types

type Update = (Maybe Vertex, Vertex -> TileFace)

An Update is a pair of a Maybe Vertex identifying the third vertex for a face addition (Nothing means it needs to be created) and a makeFace function to create the new face when given a third vertex.

type UpdateMap = Mapping DEdge Update

UpdateMap: partial map associating updates with (some) boundary directed edges.

data ForceState

ForceState: The force state records information between executing single face updates during forcing (a Boundary and an UpdateMap).

Constructors

ForceState

Fields

boundaryState :: Boundary

updateMap :: UpdateMap

type UpdateGenerator = Boundary -> [DEdge] -> UpdateMap -> UpdateMap

UpdateGenerator is the type of functions which can change the UpdateMap when given a Boundary and a focus list of particular directed boundary edges. It will add particular updates for those focus edges which match a rule. So such functions implement the forcing rules.

data BoundaryChange

BoundaryChange records a new boundary after an update plus edges which are no longer on the boundary plus a list of boundary edges revised (used to focus where the update map needs to be recalculated) (see affectedBoundary)

Constructors

BoundaryChange

Fields

newBoundary :: Boundary

removedEdges :: [DEdge]

revisedEdges :: [DEdge]

affectedBoundary :: Boundary -> [DEdge] -> [DEdge]

Given a Boundary with a list of boundary edges that have been newly added, it creates a list of boundary edges that are affected. Namely, the boundary edges sharing a vertex with a new edge. For a safe update this will be the single new edge + edges either side on the boundary. For an unsafe update this will be 4 edges including the 2 new ones. (Used to make revisedEdges in a BoundaryChange)

forcing operations

force :: Tgraph -> Tgraph

The main force function using allUGenerator representing all 10 rules

wholeTiles :: Tgraph -> Tgraph

special case of forcing only half tiles to whole tiles

forceWith :: UpdateGenerator -> Tgraph -> Tgraph

forceWith uGen: initialises force state before forcing (both using uGen to generate updates). It recursively does all updates using forceAll uGen, then gets boundary from the final state and converts back to a Tgraph

forceAll :: UpdateGenerator -> ForceState -> ForceState

forceAll uGen recursively does updates using uGen until there are no more updates,

initForceState :: UpdateGenerator -> Tgraph -> ForceState

initForceState uGen g calculates an initial force state with boundary information from g and uses uGen on all boundary edges to initialise updateMap.

reviseUpdates :: UpdateGenerator -> BoundaryChange -> UpdateMap -> UpdateMap

reviseUpdates uGen bdChange: updates the UpdateMap after boundary change (bdChange) using uGen to calculate new updates.

isSafeUpdate :: Update -> Bool

safe updates are those which do not require a new vertex, so have an identified existing vertex (Just v)

findSafeUpdate :: UpdateMap -> Maybe Update

finds the first safe update - Nothing if there are none (ordering is directed edge key ordering)

Inspecting Force Steps

stepForce :: Int -> Tgraph -> ForceState

stepForce produces an intermediate state after a given number of steps (face additions)

stepForceWith :: UpdateGenerator -> Int -> ForceState -> ForceState

do a number of force steps using a given UpdateGenerator (used by stepForce)

Single Force Steps

oneStepWith :: UpdateGenerator -> ForceState -> Maybe (ForceState, BoundaryChange)

oneStepWith uGen fs uses uGen to find updates and does one force step. It returns a (maybe) new force sate but paired with the boundary change for debugging purposes

oneStepF :: ForceState -> Maybe (ForceState, BoundaryChange)

oneStepF is a special case of oneStepWith only used for debugging

tryUnsafes :: ForceState -> Maybe BoundaryChange

tryUnsafes: Should only be used when there are no Safe updates in the UpdateMap. When touchChecKOn is True any unsafe update producing a touching vertex returns Nothing. tryUnsafes works through the unsafe updates in (directed edge) key order and completes the first unsafe update that is not blocked, returning Just the resulting boundary change (if there was one) and Nothing if all unsafes are blocked.

doSafeUpdate :: Boundary -> Update -> BoundaryChange

doSafeUpdate bd u adds a new face by completing a safe update u on boundary bd (raising an error if u is an unsafe update). It returns a BoundaryChange (containing a new Boundary, removed boundary edges and revised boundary edge list). It checks that the new face is not in conflict with existing faces, raising an error if there is a conflict.

commonV :: (Eq b, Show b) => [(b, b)] -> b

given 2 adjacent directed edges, this returns the common vertex. Raises an error if the argument is not 2 adjacent directed edges

tryUnsafeUpdate :: Boundary -> Update -> Maybe BoundaryChange

tryUnsafeUpdate bd u, calculates the resulting boundary change for an unsafe update (u) with a new vertex (raising an error if u is a safe update). It checks that the new face is not in conflict with existing faces (raising an error if there is a conflict). If touchCheckOn is True, it performs a touching vertex check with the new vertex returning Nothing if there is a touching vertex. Otherwise it returns Just a boundary change.

doUpdate :: Boundary -> Update -> BoundaryChange

doUpdate: do a single update (safe or unsafe)

Conflict Test

noConflict :: TileFace -> [TileFace] -> Bool

noConflict fc fcs where fc is a new face and fcs are neighbouring faces. There is no conflict if none of the new directed face edges of fc are already directed edges of neighbouring faces fcs (in the same direction) and the edge length types (phi/nonPhi) do not conflict.

Forcing Rules become Update Generators

Forcing rules

(wholeTileUpdates) When a join edge is on the boundary - add the missing half tile to make a whole tile.

(kiteBelowDartUpdates) When a half dart has its short edge on the boundary add the half kite that must be on the short edge (this is at ace vertices but also helps with jack and deuce vertices).

(kiteWingDartOriginUpdates) When a vertex is both a dart origin and a kite wing it must be a queen or king vertex. If there is a boundary short edge of a kite half at the vertex, add another kite half sharing the short edge. (This converts 1 kite to 2 and 3 kites to 4 in combination with the first rule).

(kiteGapUpdates) When two half kites share a short edge their oppV vertex must be a deuce vertex. Add any missing half darts needed to complete the vertex. (In the gap between the other two short edges of the full kites).

(secondTouchingDartUpdates) When a single dart wing is at a vertex which is recognised as an incomplete jack vertex and has a complete kite below the dart wing, add a second dart half touching at the vertex (sharing the kite below). This is also known as a *largeDartBase* vertex (= new dart base next level up - see later)

(sunStarUpdates) When a vertex has 3 or 4 whole kite origins (= 6 or 8 half kite origins) it must be a sun centre. Also if a vertex has 4 whole dart origins (= 8 half dart origins) it must be a star centre. Add an appropriate half kite/dart on a boundary long edge at the vertex. (This will complete suns (resp. stars) along with rule 1),

(dartKiteTopUpdates) When a dart half has its wing recognised as a jack (largeDartBase) vertex add a missing kite half on its long edge.

(thirdDartUpdates) When a vertex is a kite wing and also an origin for exactly 4 dart halves it must be a king vertex. Add a missing dart half (on any boundary long edge of a dart at the vertex).

(queenDartUpdates) If there are 4 kite wings at a vertex (necessarily a queen) add any missing half dart on a boundary kite long edge

(queenKiteUpdates) If there are 3 kite wings at a vertex (necessarily a queen) add any missing fourth half kite on a boundary kite short edge

allUGenerator :: UpdateGenerator

allUGenerator combines all the 10 update generators. They are composed (keeping the order) after applying each to the supplied boundary and focus edge list arguments

type UFinder = Boundary -> [DEdge] -> [(DEdge, TileFace)]

UFinder (Update case finder function). Given a Boundary and a list of (focus) boundary directed edges, such a function returns each focus edge satisfying the particular update case paired with the tileface matching that edge. For example, if the function is looking for dart short edges on the boundary, it will return only those focus edges which have a matching half-dart short edge, each paired with the corresponding half-dart.

type UMaker = Boundary -> TileFace -> Update

UMaker (Update creator function). Given a Boundary and a particular tileface (on the boundary), such a function produces a unique update for that edge. For example, if the function is looking to add half-kites to dart short edges it will produce the appropriate update when given a Boundary and half-dart tileface (whose short edge must be on the boundary).

boundaryFilter :: (Boundary -> DEdge -> TileFace -> Bool) -> UFinder

This is a general purpose filter used to create UFinder functions for each force rule. It requires a face predicate. The face predicate takes a Boundary bd, a boundary DEdge (a,b) and a TileFace at a (the first vertex of the directed edge) and decides whether the face is wanted or not (True = wanted) This will be used to filter all the faces at the focus edges (when given a Boundary and list of focus edges). For some predicates the boundary argument is not used (eg boundaryJoin in incompleteHalves), but for others it is used to look at all the faces at b or at other faces at a besides the supplied fc (eg kiteWDO in kitesWingDartOrigin)

makeGenerator :: UMaker -> UFinder -> UpdateGenerator

makeGenerator combines an update creator (UMaker) with its corresponding update case finder (UFinder) to create an update generator function. This is used to make all of the 10 update generators corresponding to 10 rules

When given a boundary, list of focus edges and an update map, the finder produces a list of new pairs of dEdge and face, the maker is used to convert the face in each pair to an update, the new (dedge,update) pairs are added to the final update map

Ten Update Generators (with corresponding Finders)

wholeTileUpdates :: UpdateGenerator

Update generator for rule (1)

incompleteHalves :: UFinder

Find faces with missing opposite face (mirror face)

kiteBelowDartUpdates :: UpdateGenerator

Update generator for rule (2)

nonKDarts :: UFinder

Find half darts with boundary short edge

kiteWingDartOriginUpdates :: UpdateGenerator

Update generator for rule (3) queen and king vertices add a missing kite half

kitesWingDartOrigin :: UFinder

Find kites with boundary short edge where the wing is also a dart origin

kiteGapUpdates :: UpdateGenerator

Update generator for rule (4) (for deuce vertices = largeKiteCentres) Kites whose short edge (b,a) matches a boundary edge (a,b) where their oppV (= a for LK and = b for RK) has 2 other kite halves sharing a shortE. These need a dart adding on the short edge.

kiteGaps :: UFinder

Find kite halves with a short edge on the boundary (a,b) where there are 2 other kite halves sharing a short edge at oppV of the kite half (a for left kite and b for right kite)

secondTouchingDartUpdates :: UpdateGenerator

Update generator for rule (5) secondTouchingDartUpdates - jack vertex add a missing second dart

noTouchingDarts :: UFinder

Find kite halves with a short edge on the boundary (a,b) where oppV is a largeDartBase vertex (oppV is a for left kite and b for right kite). The function mustbeLDB determines if a vertex must be a a largeDartBase

mustbeLDB :: Boundary -> Vertex -> Bool

mustbeLDB (large dart base / jack) is true of a boundary vertex if it is the wing of two darts not sharing a long edge or it is a wing of a dart and also a kite origin (false means it is either undetermined or is a large kite centre)

hasAnyMatchingE :: [DEdge] -> Bool

hasMatching asks if a directed edge list has any two matching (=opposing) directed edges.

sunStarUpdates :: UpdateGenerator

Update generator for rule (6) sunStarUpdates is for vertices that must be either sun or star almostSunStar finds half-kites/half-darts with a long edge on the boundary where their origin vertex has 8 total half-kites/half-darts respectively or their origin vertex has 6 total half-kites in the case of kites only completeSunStar will add a new face of the same type (dart/kite) sharing the long edge.

almostSunStar :: UFinder

Find a boundary long edge of either a dart where there are at least 7 dart origins, or a kite where there are at least 6 kite origins

dartKiteTopUpdates :: UpdateGenerator

Update generator for rule (7) jack vertices (largeDartBases) with dart long edge on boundary - add missing kite top

noKiteTopDarts :: UFinder

Find jack vertices (largeDartBases) with dart long edge on the boundary

thirdDartUpdates :: UpdateGenerator

Update generator for rule (8) king vertices with 2 of the 3 darts - add another half dart on a boundary long edge of existing darts

missingThirdDarts :: UFinder

Find king vertices with 2 of the 3 darts (a kite wing and 4 dart origins present) and a kite long edge on the boundary

queenDartUpdates :: UpdateGenerator

Update generator for rule (9) queen vertices (with 4 kite wings) -- add any missing half dart on a boundary kite long edge

queenMissingDarts :: UFinder

Find queen vertices (with 4 kite wings) and a boundary kite long edge

queenKiteUpdates :: UpdateGenerator

Update generator for rule (10) queen vertices with 3 kite wings -- add missing fourth half kite on a boundary kite short edge

queenMissingKite :: UFinder

Find queen vertices with only 3 kite wings and a kite short edge on the boundary

Six Update Makers

completeHalf :: UMaker

completeHalf will make an update to add a symmetric (mirror) face for a given face at a boundary join edge.

addKiteShortE :: UMaker

add a (missing) half kite on a (boundary) short edge of a dart or kite

addDartShortE :: UMaker

add a half dart top to a boundary short edge of a half kite.

completeSunStar :: UMaker

add a kite half to a kite long edge or dart half to a dart long edge

addKiteLongE :: UMaker

add a kite to a long edge of a dart or kite

addDartLongE :: UMaker

add a half dart on a boundary long edge of a dart or kite

Other Forcing operations

forceSub :: SubTgraph -> SubTgraph

force applied to a SubTgraph - has no effect on tracked subsets but applies force to the full Tgraph.

addHalfKite :: Tgraph -> DEdge -> Tgraph

[addHalfKite and addHalfDart are not efficient but used to tinker with a Tgraph by adding a single half tile e.g. to see how it affects forcing. They use the update makers (UMaker).]

addHalfKite is for adding a single half kite on a chosen DEdge of a Tgraph. The DEdge must be a boundary edge but the direction is not important as the correct direction is automatically calculated. It will fail if the edge is a dart join.

addHalfDart :: Tgraph -> DEdge -> Tgraph

addHalfDart is for adding a single half dart on a chosen DEdge of a Tgraph. The DEdge must be a boundary edge but the direction is not important as the correct direction is automatically calculated. It will fail if the edge is a dart short edge or kite join.

forceLDB :: Vertex -> Tgraph -> Tgraph

For an unclassifiable dart wing v in a Tgraph, force it to become a large dart base (largeDartBase) by adding a second half dart face (sharing the kite below the existing half dart face at v) This assumes exactly one dart wing tip is at v, and that half dart has a full kite below it.

forceLKC :: Vertex -> Tgraph -> Tgraph

For an unclassifiable dart wing v in a Tgraph, force it to become a large kite centre (largeKiteCentres) by adding 3 faces - a second half dart face sharing the long edge of the existing half dart face at v, and then completing the kite on the new half dart short edge. This assumes exactly one dart wing tip is at v. (Note: farK for a half-dart d is that half of a full kite attached to the short edge of d which does not share an edge with d). It is safe to add the 3 parts because v being unknown ensures the existing dart has a boundary long edge and the new farK does not already exist (attached to existing dart farK), provided the existing dart half has no kite or a full kite below. If it has only a half kite below, but the new farK exists, then v will already be a crossing boundary.

mustFind :: Foldable t => (p -> Bool) -> t p -> p -> p

mustFind is used to search (in forceLKC and forceLDB) mustFind p ls err returns the first item in ls satisfying predicate p and returns err argument when none found

Adding faces (Auxiliary operations)

findThirdV :: Boundary -> DEdge -> (Int, Int) -> Maybe Vertex

findThirdV is the main exported interface to find a third vertex for a face added to the RIGHT HAND SIDE of a directed boundary edge. It is a simpler interface for searchThirdV (which it uses). In findThirdV bd (a,b) (n,m), the two integer arguments n and m are the INTERNAL angles for the new face on the boundary edge (a,b) (for a and b respectively) expressed as multiples of tt (tt being a tenth turn) and must both be either 1,2, or 3 It converts n and m to the correct IntAngles, subtracting n from 10 to get the antiClockwise (external) angle on the first vertex (a), before calling searchThirdV to return a Maybe Vertex. (See also searchThirdV)

anglesForJoinRD :: (Int, Int)

mnemonic for internal angles of an edge (expressed as integer units of a tenth turn (I.e 1,2 or 3)

anglesForJoinLD :: (Int, Int)

mnemonic for internal angles of an edge (expressed as integer units of a tenth turn (I.e 1,2 or 3)

anglesForJoinRK :: (Int, Int)

mnemonic for internal angles of an edge (expressed as integer units of a tenth turn (I.e 1,2 or 3)

anglesForJoinLK :: (Int, Int)

mnemonic for internal angles of an edge (expressed as integer units of a tenth turn (I.e 1,2 or 3)

anglesForLongLD :: (Int, Int)

mnemonic for internal angles of an edge (expressed as integer units of a tenth turn (I.e 1,2 or 3)

anglesForLongRD :: (Int, Int)

mnemonic for internal angles of an edge (expressed as integer units of a tenth turn (I.e 1,2 or 3)

anglesForLongRK :: (Int, Int)

mnemonic for internal angles of an edge (expressed as integer units of a tenth turn (I.e 1,2 or 3)

anglesForLongLK :: (Int, Int)

mnemonic for internal angles of an edge (expressed as integer units of a tenth turn (I.e 1,2 or 3)

anglesForShortLD :: (Int, Int)

mnemonic for internal angles of an edge (expressed as integer units of a tenth turn (I.e 1,2 or 3)

anglesForShortRD :: (Int, Int)

mnemonic for internal angles of an edge (expressed as integer units of a tenth turn (I.e 1,2 or 3)

anglesForShortLK :: (Int, Int)

mnemonic for internal angles of an edge (expressed as integer units of a tenth turn (I.e 1,2 or 3)

anglesForShortRK :: (Int, Int)

mnemonic for internal angles of an edge (expressed as integer units of a tenth turn (I.e 1,2 or 3)

newtype IntAngle

IntAngles are Ints mod 10

Constructors

IntAngle Int

intAngle :: Int -> IntAngle

make an IntAngle from an Int (using mod 10)

anti :: IntAngle -> IntAngle -> IntAngle

from IntAngle n, turn clockwise/anticlockwise by IntAngle m to get a new IntAngle

clock :: IntAngle -> IntAngle -> IntAngle

from IntAngle n, turn clockwise/anticlockwise by IntAngle m to get a new IntAngle

searchThirdV :: Boundary -> DEdge -> IntAngle -> IntAngle -> Maybe Vertex

searchThirdV is the main function for constructing a new face. searchThirdV bd (a,b) nAnti nClock will find the third vertex of a face to be added to the RIGHT HAND SIDE of boundary directed edge (a,b) by inspecting the edges found at a and b. (To add to the LHS of (a,b) then the boundary direction must be (b,a) so add to the right of (b,a))

nAnti is the IntAngle searched for at vertex a going anticlockwise (starting with (a,b) as IntAngle 0). nAnti should be formed from integers 7, 8, or 9 for the face to be on RHS.

nClock is the IntAngle searched for at vertex b going clockwise (starting with (b,a) as IntAngle 0). nClock should be formed from integers 1,2, or 3 for the face to be on RHS.

If existing edges are found in the specified directions they are used and the associated third vertex is returned (Just v). If nothing is found, Nothing is returned - indicating that a new vertex is needed.

A gap in faces processed round a and b (i.e. a boundary edge before all faces are checked) will indicate a *crossing boundary* error where it is unsafe to conclude the edges do not exist. This problem should be detected by findClockAt and findAntiAt.

findClockAt :: Boundary -> IntAngle -> (Vertex, Vertex) -> [TileFace] -> Maybe Vertex

findClockAt bd n (a,b) fcs (for a boundary bd and boundary directed edge (a,b) and faces at b fcs) findClockAt bd n (a,b) fcs will look at directed edges from b of fcs going anti-clockwise round b with their direction starting with (b,a) in direction 0. It then returns Just c if (b,c) is found in direction n on the boundary. It will return Nothing if no such edge (b,c) is found BUT a *crossing boundary* error if a boundary edge is found before the sought angle. It also produces an error if the last angle found is below the one searched for. This indicates an incorrect graph, and the boundary argument is used just for reporting the error in such cases.

findClockAt, findAntiAt For a boundary bd and boundary directed edge (a,b) findAntiAt bd n (a,b) fcs will look at directed edges from a of fcs going anti-clockwise round a with their direction starting with (a,b) in direction 0. fcs must be the faces at a. It then returns Just c if (a,c) is found in direction n on the boundary. It will return Nothing if no such edge (a,c) is found BUT a *crossing boundary* error if a boundary edge is found before the sought angle. It also produces an error if the last angle found is beyond the one searched for. This indicates an incorrect graph, and the boundary argument is used just for reporting the error in such cases. Similarly findClockAt in the clockwise direction but called with the reverse directed edge, so fcs must be faces at b.

findAntiAt :: Boundary -> IntAngle -> (Vertex, Vertex) -> [TileFace] -> Maybe Vertex

findAntiAt bd n (a,b) fcs (for a boundary bd and boundary directed edge (a,b) and faces at a fcs) will look at directed edges from a of fcs going anti-clockwise round a with their direction starting with (a,b) in direction 0. It then returns Just c if (a,c) is found in direction n on the boundary. It will return Nothing if no such edge (a,c) is found BUT a *crossing boundary* error if a boundary edge is found before the sought angle. It also produces an error if the last angle found is beyond the one searched for. This indicates an incorrect graph, and the boundary argument is used just for reporting the error in such cases.

findClockAt, findAntiAt For a boundary bd and boundary directed edge (a,b) findAntiAt bd n (a,b) fcs will look at directed edges from a of fcs going anti-clockwise round a with their direction starting with (a,b) in direction 0. fcs must be the faces at a. It then returns Just c if (a,c) is found in direction n on the boundary. It will return Nothing if no such edge (a,c) is found BUT a *crossing boundary* error if a boundary edge is found before the sought angle. It also produces an error if the last angle found is beyond the one searched for. This indicates an incorrect graph, and the boundary argument is used just for reporting the error in such cases. Similarly findClockAt in the clockwise direction but called with the reverse directed edge, so fcs must be faces at b.

errorCheckAnti :: Boundary -> Vertex -> IntAngle -> [(IntAngle, Vertex)] -> Maybe Vertex

errorCheckAnti bd v n and errorCheckClock bd v n are used instead of just lookup n, because the search for n should find it in the front pair on the list or not at all. However if the first angle is smaller than n for checkClockFor or bigger than n for checkAntiFor, there is an incorrect graph. The first 2 arguments are passed only to report such errors.

errorCheckClock :: Boundary -> Vertex -> IntAngle -> [(IntAngle, Vertex)] -> Maybe Vertex

errorCheckAnti bd v n and errorCheckClock bd v n are used instead of just lookup n, because the search for n should find it in the front pair on the list or not at all. However if the first angle is smaller than n for checkClockFor or bigger than n for checkAntiFor, there is an incorrect graph. The first 2 arguments are passed only to report such errors.

smallerAngle :: IntAngle -> IntAngle -> Bool

x is a smaller IntAngle than y if the (mod 10) integer of x is smaller than the (mod 10) integer of y

allAnglesClock :: [(IntAngle, Vertex)] -> [TileFace] -> [(IntAngle, Vertex)]

allAnglesClock and allAnglesAnti are used by findClockAt and findAntiAt. The second argument is the unprocessed list of faces attached to a common vertex a. The first argument is the accumulated list of directions found so far with last one at the front of the list e.g (n,last) means the last edge found was (a,last) at angle n. It looks for another face sharing vertex last (and therefore edge (a,last)) and uses this to get the next angle and edge. The found face is removed from the list of faces to be processed in the recursive call. If there is no such face, it will return the whole (angle,vertex) list found so far provided there are no faces left over. Faces left over indicate a crossing boundary at the vertex and therefore an error should be reported (because the angle list is incomplete and unsafe to use).

allAnglesAnti :: [(IntAngle, Vertex)] -> [TileFace] -> [(IntAngle, Vertex)]

allAnglesClock and allAnglesAnti are used by findClockAt and findAntiAt. The second argument is the unprocessed list of faces attached to a common vertex a. The first argument is the accumulated list of directions found so far with last one at the front of the list e.g (n,last) means the last edge found was (a,last) at angle n. It looks for another face sharing vertex last (and therefore edge (a,last)) and uses this to get the next angle and edge. The found face is removed from the list of faces to be processed in the recursive call. If there is no such face, it will return the whole (angle,vertex) list found so far provided there are no faces left over. Faces left over indicate a crossing boundary at the vertex and therefore an error should be reported (because the angle list is incomplete and unsafe to use).

intAngleAt :: Vertex -> TileFace -> IntAngle

intAngleAt v fc gives the internal angle of the face fc at vertex v (which must be a vertex of the face) returning an IntAngle (1,2,or 3).

faceAngles :: TileFace -> [IntAngle]

faceAngles returns a list of the three internal angles of a face (clockwise from originV) represented as IntAngles - always 1 or 2 for kites and 1 or 3 for darts

testAngles :: Tgraph -> (Vertex, Vertex) -> ([(IntAngle, Vertex)], [(IntAngle, Vertex)])

For testing allAnglesAnti and allAnglesClock on a boundary directed edge. The RHS of directed edge (a,b) should be the exterior side.