Tgraph.Prelude

Description

Introduces Tgraphs and includes operations on vertices, edges and faces as well as Tgraphs. Also includes functions to calculate (relative) locations of vertices (createVPoints) and re-exports module HalfTile

Documentation

type Mapping = Map

Mapping abbreviates Data.Map.Map

Tgraphs

type Vertex = Int

Tgraph vertices

type DEdge = (Vertex, Vertex)

directed edge

type TileFace = HalfTile (Vertex, Vertex, Vertex)

Tgraph faces (vertices clockwise starting with tile origin vertex) a specialisation of HalfTile

data Tgraph

A Tgraph contains vertices, and faces (each are lists treated as sets with no repetitions). Every vertex must be a face vertex and vice versa

Constructors

Tgraph

Fields

vertices :: [Vertex]

faces :: [TileFace]

makeTgraph :: [TileFace] -> Tgraph

provided a list of faces makes sense and are face-edge-connected makeTgraph will create a Tgraph from the faces by calculating vertices

checkTgraph :: [TileFace] -> Tgraph

checkTgraph creates a graph from faces but checks for edge conflicts and crossing boundaries and connectedness No crossing boundaries and connected => face-connected

selectFaces :: [TileFace] -> Tgraph -> Tgraph

select or remove faces from a Tgraph, but check resulting graph for connectedness and no crossing boundaries

removeFaces :: [TileFace] -> Tgraph -> Tgraph

select or remove faces from a Tgraph, but check resulting graph for connectedness and no crossing boundaries

nullGraph :: Tgraph -> Bool

is the graph empty?

Tests and Tgraph properties

conflictingDedges :: Tgraph -> [DEdge]

conflictingDedges g returns a list of conflicting directed edges of the faces in g (which should be null)

conflictingLengthEdges :: Tgraph -> [DEdge]

conflictingLengthEdges g returns a list of conflicting lengthed edges of the faces in g (which should be null)

edgeConflicts :: Tgraph -> Bool

predicate - true if there are edge conflicts in a Tgraph

crossingBVs :: Tgraph -> [Vertex]

crossingBVs g returns a list of vertices with crossing boundaries (which should be null).

crossingBoundaries :: Tgraph -> Bool

There are crossing boundaries if vertices occur more than once at the start of all boundary directed edges (or more than once at the end of all boundary directed edges).

connected :: Tgraph -> Bool

predicate to check Tgraph is a connected graph

connectedTo :: Eq a => a -> [a] -> [(a, a)] -> [a]

auxiliary function for calculating connectedness by depth first search connectedTo v unvisited edges returns list of vertices connected to v from the list of vertices (unvisited) using given list of edges

Face and Vertex Operations

ldarts :: Tgraph -> [TileFace]

selecting left darts, right darts, left kite, right kites from a Tgraph

rdarts :: Tgraph -> [TileFace]

selecting left darts, right darts, left kite, right kites from a Tgraph

lkites :: Tgraph -> [TileFace]

selecting left darts, right darts, left kite, right kites from a Tgraph

rkites :: Tgraph -> [TileFace]

selecting left darts, right darts, left kite, right kites from a Tgraph

valencyD :: Tgraph -> Vertex -> Int

directed edge valency of a vertex

faceVs :: TileFace -> (Vertex, Vertex, Vertex)

triple of face vertices in order clockwise - tileRep specialised to TileFace

faceVList :: TileFace -> [Vertex]

list of (three) face vertices in order clockwise

firstV :: TileFace -> Vertex

Whilst first, second and third vertex of a face are obvious (clockwise), it is often more convenient to refer to the originV (=firstV), oppV (the vertex at the other end of the join edge), and wingV (the remaining vertex not on the join edge)

secondV :: TileFace -> Vertex

Whilst first, second and third vertex of a face are obvious (clockwise), it is often more convenient to refer to the originV (=firstV), oppV (the vertex at the other end of the join edge), and wingV (the remaining vertex not on the join edge)

thirdV :: TileFace -> Vertex

Whilst first, second and third vertex of a face are obvious (clockwise), it is often more convenient to refer to the originV (=firstV), oppV (the vertex at the other end of the join edge), and wingV (the remaining vertex not on the join edge)

originV :: TileFace -> Vertex

Whilst first, second and third vertex of a face are obvious (clockwise), it is often more convenient to refer to the originV (=firstV), oppV (the vertex at the other end of the join edge), and wingV (the remaining vertex not on the join edge)

wingV :: TileFace -> Vertex

Whilst first, second and third vertex of a face are obvious (clockwise), it is often more convenient to refer to the originV (=firstV), oppV (the vertex at the other end of the join edge), and wingV (the remaining vertex not on the join edge)

oppV :: TileFace -> Vertex

Whilst first, second and third vertex of a face are obvious (clockwise), it is often more convenient to refer to the originV (=firstV), oppV (the vertex at the other end of the join edge), and wingV (the remaining vertex not on the join edge)

indexV :: Int -> TileFace -> Int

indexV finds the index of a vertex in a face (firstV -> 0, secondV -> 1, thirdV -> 2)

nextV :: Vertex -> TileFace -> Vertex

nextV returns the next vertex in a face going clockwise from v where v must be a vertex of the face

prevV :: Vertex -> TileFace -> Vertex

prevV returns the previous vertex in a face (i.e. next going anti-clockwise) from v where v must be a vertex of the face

isAtV :: Vertex -> TileFace -> Bool

isAtV v fc asks if a face fc has v as a vertex

makeNewVs :: Int -> [Vertex] -> [Vertex]

given existing vertices vs, create n new vertices

makeNewV :: [Vertex] -> Vertex

return one new vertex

Edge Operations

faceDedges :: TileFace -> [DEdge]

directed edges (clockwise) round a face

reverseE :: DEdge -> DEdge

opposite directed edge

firstE :: TileFace -> DEdge

Whilst first, second and third edges are obvious (always clockwise), it is often more convenient to refer to the joinE (join edge), shortE (the short edge which is not a join edge), and longE (the long edge which is not a join edge) these are also directed clockwise. joinOfTile also returns the join edge but in the direction away from the origin

secondE :: TileFace -> DEdge

Whilst first, second and third edges are obvious (always clockwise), it is often more convenient to refer to the joinE (join edge), shortE (the short edge which is not a join edge), and longE (the long edge which is not a join edge) these are also directed clockwise. joinOfTile also returns the join edge but in the direction away from the origin

thirdE :: TileFace -> DEdge

Whilst first, second and third edges are obvious (always clockwise), it is often more convenient to refer to the joinE (join edge), shortE (the short edge which is not a join edge), and longE (the long edge which is not a join edge) these are also directed clockwise. joinOfTile also returns the join edge but in the direction away from the origin

joinE :: TileFace -> DEdge

joinE preserves the clockwise direction unlike joinOfTile

Whilst first, second and third edges are obvious (always clockwise), it is often more convenient to refer to the joinE (join edge), shortE (the short edge which is not a join edge), and longE (the long edge which is not a join edge) these are also directed clockwise. joinOfTile also returns the join edge but in the direction away from the origin

shortE :: TileFace -> DEdge

shortE not the join edge in the dart cases

Whilst first, second and third edges are obvious (always clockwise), it is often more convenient to refer to the joinE (join edge), shortE (the short edge which is not a join edge), and longE (the long edge which is not a join edge) these are also directed clockwise. joinOfTile also returns the join edge but in the direction away from the origin

longE :: TileFace -> DEdge

longE not the join edge in the kite cases

Whilst first, second and third edges are obvious (always clockwise), it is often more convenient to refer to the joinE (join edge), shortE (the short edge which is not a join edge), and longE (the long edge which is not a join edge) these are also directed clockwise. joinOfTile also returns the join edge but in the direction away from the origin

joinOfTile :: TileFace -> DEdge

The directed join edge of a face but origin first (not clockwise for RD and LK)

Whilst first, second and third edges are obvious (always clockwise), it is often more convenient to refer to the joinE (join edge), shortE (the short edge which is not a join edge), and longE (the long edge which is not a join edge) these are also directed clockwise. joinOfTile also returns the join edge but in the direction away from the origin

facePhiEdges :: TileFace -> [DEdge]

The phi edges of a face (both directions) which is long edges of darts, and join and long edges of kites

faceNonPhiEdges :: TileFace -> [DEdge]

The non-phi edges of a face (both directions) which is short edges of kites, and join and short edges of darts

matchingE :: (TileFace -> DEdge) -> TileFace -> TileFace -> Bool

matchingE etype fc is a predicate on tile faces where etype finds a particular type of edge of a face etype could be joinE or longE or shortE for example it maps fc' to True if fc' has an etype edge matching the (reversed) etype edge of fc

matchingLongE :: TileFace -> TileFace -> Bool

special cases of matchingE etype where etype is longE, shortE and joinE

matchingShortE :: TileFace -> TileFace -> Bool

special cases of matchingE etype where etype is longE, shortE and joinE

matchingJoinE :: TileFace -> TileFace -> Bool

special cases of matchingE etype where etype is longE, shortE and joinE

graphDedges :: Tgraph -> [(Vertex, Vertex)]

all the directed edges of a graph

phiEdges :: Tgraph -> [(Vertex, Vertex)]

phiEdges returns a list of the longer edges of a Tgraph (both directions of each edge)

nonPhiEdges :: Tgraph -> [(Vertex, Vertex)]

nonPhiEdges returns a list of the shorter edges of a Tgraph (both directions of each edge)

graphEdges :: Tgraph -> [(Vertex, Vertex)]

graphEdges returns a list of all the edges of a Tgraph (both directions of each edge)

bothDir :: [DEdge] -> [DEdge]

bothDir adds missing reverse directed edges to a list of directed edges and then removes duplicates

bothDir' :: [DEdge] -> [DEdge]

bothDir' adds the reverse directed edges to a list of directed edges without checking for duplicates

boundaryDedges :: Tgraph -> [(Vertex, Vertex)]

boundaryDedges g are missing reverse directed edges in graphDedges g (these are single directions only) Direction is such that a face is on LHS and exterior is on RHS of each boundary directed edge

boundaryEdges :: Tgraph -> [(Vertex, Vertex)]

boundary edges are face edges not shared by 2 faces (both directions)

internalEdges :: Tgraph -> [(Vertex, Vertex)]

internal edges are shared by two faces = all edges except boundary edges

edgeNb :: TileFace -> TileFace -> Bool

two tile faces are edge neighbours

Vertex Location Calculation

createVPoints :: [TileFace] -> Mapping Vertex (Point V2 Double)

createVPoints: process list of faces to associate points for each vertex. Faces must be tile-connected.

addVPoints :: [TileFace] -> [TileFace] -> Mapping Vertex (Point V2 Double) -> Mapping Vertex (Point V2 Double)

addVPoints readyfaces fcOther vpMap The first argument list of faces (readyfaces) contains the ones being processed next in order where each will have at least two known vertex points. The second argument list of faces (fcOther) have not yet been added and may not yet have known vertex points. The third argument is the mapping of vertices to points. This is used in tryUpdate as well as createVPoints

initJoin :: TileFace -> Mapping Vertex (Point V2 Double)

initJoin fc initialises a vpMap with locations for join edge vertices of fc along x axis - used to initialise createVPoints

find3Locs :: (Vertex, Vertex, Vertex) -> Mapping Vertex (Point V2 Double) -> (Maybe (Point V2 Double), Maybe (Point V2 Double), Maybe (Point V2 Double))

lookup 3 vertex locations

thirdVertexLoc :: TileFace -> Mapping Vertex (Point V2 Double) -> Maybe (Vertex, Point V2 Double)

thirdVertexLoc fc vpMap

New Version - Assumes all edge lengths are 1 or phi It now uses signorm to produce vectors of length 1 rather than rely on relative lengths.

Requires ttangle and phi from TileLib

thirdVertexLoc fc vpMap where fc is a tileface and vpMap associates points with vertices (positions) It looks up all 3 vertices in vpMap hoping to find 2 of them, it then returns Just pr where pr is an association pair for the third vertex. If all 3 are found, returns Nothing If none or one found this is an error (a non tile-connected face)

SubTgraphs

data SubTgraph

SubTgraph - introduced to allow tracking of subsets of faces in both force and decompose oerations A SubTgraph has a main Tgraph (fullgraph) and a list of subsets of faces. The list allows for tracking different subsets of faces at the same time

Constructors

SubTgraph

Fields

fullGraph :: Tgraph

trackedSubsets :: [[TileFace]]

makeSubTgraph :: Tgraph -> [[TileFace]] -> SubTgraph

makeSubTgraph g trackedlist creates a SubTgraph from a Tgraph g from trackedlist where each list in trackedlist is a subsets of the faces of g (any faces not in g are ignored)