Tgraph.Relabelling

Description

This module includes relabelling functions for Tgraphs whose main purpose is to implement a guided union of Tgraphs (simpleUnion and trySimpleUnion)

Relabelling operations

type Relabelling = IntMap Vertex

Relabelling is a special case of mappings from vertices to vertices. We use the identity map for vertices not found in the mapping domain (see relabelV). Relabellings are expected to be 1-1 on their domain.

relabelGraph :: Relabelling -> Tgraph -> Tgraph

Uses a relabelling to change vertices in a Tgraph. relabelGraph rlab g will produce a valid Tgraph provided: g is a valid Tgraph, and the mapping rlab extended with the identity is 1-1 on vertices in g. (Vertices of g that are not in the domain of the mapping are left unchanged.)

relabelFace :: Relabelling -> TileFace -> TileFace

Uses a relabelling to relabel the three vertices of a face. Any vertex not in the domain of the mapping is left unchanged. The mapping should be 1-1 on the 3 vertices to avoid creating a self loop edge.

relabelV :: Relabelling -> Vertex -> Vertex

relabelV rlab v. Uses relabelling rlab to find a replacement for v (leaves as v if none found). I.e relabelV turns a relabelling into a total function using identity for undefined cases.

relabelAvoid :: [Vertex] -> Tgraph -> Tgraph

relabelAvoid vs g - produces a new Tgraph from g by relabelling. Any vertex in g that is in the list vs will be changed to a new vertex that is neither in g nor in the list vs. Vertices in g that are not in vs will remain the same.

prepareFixAvoid :: [Vertex] -> [Vertex] -> Tgraph -> Tgraph

prepareFixAvoid fix avoid g - produces a new Tgraph from g by relabelling. Any vertex in g that is in the list avoid but not in the list fix will be changed to a new vertex that is neither in g nor in the list avoid \\ fix. All other vertices of g (including those in fix) will remain the same. Usage: This is used to prepare a graph by avoiding accidental label clashes with the avoid list (usually vertices of another graph). However we fix a list of vertices which we intend to control in a subsequent relabelling. (this is usually a pair of vertices from a directed edge that will get a specific subsequent relabelling).

relabelAny :: Tgraph -> Tgraph

Relabel all vertices in a Tgraph using new labels 1..n (where n is the number of vertices).

Union (and matching) operations

simpleUnion :: (Tgraph, DEdge) -> (Tgraph, DEdge) -> Tgraph

simpleUnion (g1,e1) (g2,e2) - where edge e1 is in g1 and e2 is in g2, checks if g2 can be relabelled to produce a common single region of overlap with g1 (with e2 relabelled to e1). If so then the result is a (checked) Tgraph g where g has the union of the faces. Otherwise an error is raised.

CAVEAT: The overlap must be a SINGLE tile-connected region in g1.

trySimpleUnion :: (Tgraph, DEdge) -> (Tgraph, DEdge) -> ReportFail Tgraph

trySimpleUnion (g1,e1) (g2,e2) - where edge e1 is in g1 and e2 is in g2, checks if g2 can be relabelled to produce a common single region of overlap with g1 (with e2 relabelled to e1). If so then the result is Right g where g is the union of the faces. Otherwise the result is Left lines where lines explains the problem.

CAVEAT: The overlap must be a SINGLE tile-connected region in g1.

fullUnion :: (Tgraph, DEdge) -> (Tgraph, DEdge) -> Tgraph

fullUnion (g1,e1) (g2,e2) will try to create the full union of g1 and g2 by matching the respective edges e1 and e2 and relabelling g2 to match g1. It will raise an error if there is a mismatch. This version uses geometry of tiles (vertex locations) to correct for multiple overlapping regions of tiles in g1 and g2

tryFullUnion :: (Tgraph, DEdge) -> (Tgraph, DEdge) -> ReportFail Tgraph

tryFullUnion (g1,e1) (g2,e2) will try to create the full union of g1 and g2 by matching the respective edges e1 and e2 and relabelling g2 to match g1. It returns Left lines if there is a mismatch (where lines explains the problem). If succesfull it returns Right g where g is the resulting union (checked for Tgraph properties) This version uses geometry of tiles (vertex locations) to correct for multiple overlapping regions of tiles in g1 and relabelled g2

matchByEdges :: (Tgraph, DEdge) -> (Tgraph, DEdge) -> Tgraph

matchByEdges (g1,e1) (g2,e2) produces a relabelled version of g2 that is consistent with g1 on their overlap. The overlapping region must contain the directed edge e1 in g1. The edge e2 in g2 will be identified with e1 by the relabelling of g2. This produces an error if a mismatcch is found in the overlap.

CAVEAT: The overlap must be a SINGLE tile-connected region in g1. (If the overlap contains more than one tile-connected region the result may not be a correct relabelling of g2)

tryMatchByEdges :: (Tgraph, DEdge) -> (Tgraph, DEdge) -> ReportFail Tgraph

tryMatchByEdges (g1,e1) (g2,e2) produces either Right g where g is a relabelled version of g2 that is consistent with g1 on their overlap or Left lines if there is a mismatch (lines explaining the problem). The overlapping region must contain the directed edge e1 in g1. The edge e2 in g2 will be identified with e1 by the relabelling of g2.

CAVEAT: The overlap must be a SINGLE tile-connected region in g1. (If the overlap contains more than one tile-connected region the result may not be a correct relabelling of g2)

matchByCommonEdge :: Tgraph -> DEdge -> Tgraph -> Tgraph

matchByCommonEdge g1 e g2 produces a relabelled version of g2 that is consistent with g1 on their overlap. The overlaping region must contain the common directed edge e without relabelling. This produces an error if a mismatcch is found in the overlap.

CAVEAT: The overlap must be a SINGLE tile-connected region in g1. (If the overlap contains more than one tile-connected region the result may not be a correct relabelling of g2)

tryMatchByCommonEdge :: Tgraph -> DEdge -> Tgraph -> ReportFail Tgraph

tryMatchByCommonEdge g1 e g2 produces either Right g where g is a relabelled version of g2 that is consistent with g1 on their overlap or Left lines if there is a mismatch (lines explaining the problem). The overlaping region must contain the common directed edge e without relabelling.

CAVEAT: The overlap must be a SINGLE tile-connected region in g1. (If the overlap contains more than one tile-connected region the result may not be a correct relabelling of g2)

Auxiliary functions

findRelabelling :: (Tgraph, TileFace) -> (Tgraph, TileFace) -> ReportFail Relabelling

findRelabelling is an auxiliary function for tryMatchByEdges and tryMatchByCommonEdge. findRelabelling (g1,fc1) (g2,fc2) - fc1 and fc2 should be matching face types, with fc1 in g1 and fc2 in g2. g2 must have no vertices in common with g1 except for (possibly) vertices in fc2. The result is either Right rel where rel is a relabelling map to make g2 consistent with g1 in a region of overlap if this is possible, or Left lines if there is a mismatch (lines explaining the problem). In the successful case rel when applied to fc2 will be identical to fc1.

CAVEAT: The common overlap needs to be a SINGLE region of tile-connected faces in g1. The common region must contain fc1 in g1 and fc2 in g2. This may not produce correct results if there are other overlaps not tile-connected in g1 to the region containing fc1

relabelTouching :: Tgraph -> Tgraph -> Tgraph

relabelTouching is used by tryFullUnion. relabelTouching g1 g2 assumes that there are vertex labels in g2 that match with vertex labels in g1 in a matching tile-connected overlap region. It then identifies other vertices in g2 that need to be relabelled to match vertices in g1 based on a vertex location calculation. It returns the result of applying this further relabelling to g2.

initRelabelling :: TileFace -> TileFace -> Relabelling

initRelabelling f1 f2 - creates a relabelling so that if applied to face f2, the vertices will match with face f1 exactly. It does not check that the tile faces have the same form (LK,RK,LD,RD).

differences :: [(Vertex, Vertex)] -> [(Vertex, Vertex)]

selects only non-matching pairs from a list

addRelabel :: Tgraph -> [TileFace] -> [TileFace] -> Relabelling -> ReportFail Relabelling

addRelabel is used by findRelabelling to build a relabelling map which can fail, producing Left lines. In the successful case (addRelabel g processing awaiting rlab) produces a Right rel where rel is the required relabelling. The arguments are: g - the Tgraph being matched against. processing - a list of faces to be matched next where each has an edge in common with at least one previously matched face (or it is the starting face). awaiting - a list of faces that have not yet been tried for a match and are not tile-connected to any faces already matched. rlab - a vertex to vertex mapping used to store relabelling information during the process.

The idea is that from a single matched starting face we process faces that share an edge with a previously matched face. We process faces that have a match in g (with 2 matching vertices). If a face is tried but has no such match, it is ignored (it may share some boundary with g, but for the overlap to be a single tile-connected region, only boundaries with matched tiles are possible and therefore relabelling will already be done for the boundary). If a processed face has an edge in common with a face in g it has to match exactly apart from (possibly) the third vertex label, otherwise the faces do not match and this indicates a mismatch on the overlap and Left ... is returned.

matchFaceIn :: Tgraph -> TileFace -> ReportFail (Maybe TileFace)

matchFaceIn g f - looks for a face in g that corresponds to f (sharing a directed edge), If the corresponding face does not match properly (with twoVMatch) this stops the matching process returning Left ... to indicate a failed match. Otherwise it returns either Right (Just f) where f is the matched face or Right Nothing if there is no corresponding face.

twoVMatch :: TileFace -> TileFace -> Bool

twoVMatch f1 f2 is True if the two tilefaces are the same except for a single vertex label possibly not matching.

sameFaceType :: HalfTile rep1 -> HalfTile rep2 -> Bool

sameFaceType f1 f2 is True if f1 and f2 are of the same form (i.e. both LD or both RD or both LK or both RK)