Tgraphs

Description

This is the main module for Tgraphs which collects and exports the various other Tgraph modules and includes a definition of emplace and other experimental combinations.

Documentation

type Relabelling = Map Vertex Vertex

Relabelling is a special case of mappings from vertices to vertices. We use the identity map for vertices not found in the mapping domain (see relabelV). Relabellings are expected to be 1-1.

emplace :: Tgraph -> Tgraph

emplace does maximal composing with force and composeG, then applies decomposeG and force repeatedly back to the starting level. It produces the emplacement of influence of the argument graph.

emplacements :: Tgraph -> [Tgraph]

emplacements is best supplied with a maximally composed or near maximally composed graph It produces an infinite list of emplacements of the starting graph and its decompositions.

emplaceChoices :: Tgraph -> [Tgraph]

a version of emplace using makeChoices at the top level. after makeChoices we use emplace to attempt further compositions but with no further choices

makeChoices :: Tgraph -> [Tgraph]

makeChoices is a temporary tool which does not attempt to analyse choices for correctness. It can thus create some choices which will be incorrect. The unknowns returned from classifyDartWings can become largeKiteCentres or largeDartBases. This produces 2^n choices where n is the number of unknowns (excluding lone dart wing tips with valencyD 2).

compForce :: Tgraph -> Tgraph

compForce does a force then composeG but it by-passes the check on the composed graph because the argument was forced

allCompFs :: Tgraph -> [Tgraph]

allCompFs g produces a list of all forced compositions starting from g up to but excluding the empty graph

allComps :: Tgraph -> [Tgraph]

allComps g produces a list of all compositions starting from g up to but excluding the empty graph. This is not safe in general as it can fail by producing a non-connected graph or graph with crossing boundaries.

maxCompose :: Tgraph -> Tgraph

maxCompose and maxFCompose produce a maximal graph.

maxFCompose :: Tgraph -> Tgraph

maxCompose and maxFCompose produce a maximal graph.

removeIncompleteTiles :: Tgraph -> Tgraph

remove halftile faces that do not have their matching half tile

relabelGraph :: Relabelling -> Tgraph -> Tgraph

Uses a relabelling to change vertices in a Tgraph relabelGraph vmap g will produce a valid Tgraph provided: g is a valid Tgraph, and the mapping vmap extended with the identity is 1-1 on vertices in g. (Vertices of g that are not in the domain of the mapping are left unchanged.)

relabelFace :: Relabelling -> TileFace -> TileFace

Uses a relabelling to relabel the three vertices of a face Any vertex not in the domain of the mapping is left unchanged. The mapping should be 1-1 on the 3 vertices to avoid creating a self loop edge.

relabelV :: Relabelling -> Vertex -> Vertex

relabelV vmap v. Uses relabelling vmap to find a replacement for v (leaves as v if none found) I.e relabelV turns a relabelling into a total function using identity for undefined cases.

all3 :: (a -> b) -> (a, a, a) -> (b, b, b)

Applies a function to all three elements of a triple

partRelabelFixChange :: [Vertex] -> [Vertex] -> Tgraph -> Tgraph

partRelabelFixChange fix change g - produces a new Tgraph from g by relabelling. Any vertex in g that is in the list fix will remain unchanged. All Other vertices in g that are in the list change will be changed to new vertices that are neither in g nor in the list change nor in the list fix. (If a vertex in g appears in both fix and change, it will remain unchanged)

newVMapAvoid :: [Vertex] -> [Vertex] -> Relabelling

newVMapAvoid avoid vs - produces a 1-1 mapping from vertices in vs to new vertices such that no new vertex is in the list avoid or in vs

relabelAny :: Tgraph -> Tgraph

Relabel all vertices in Tgraph using new labels 1..n (where n is the number of vertices)

relabelByCommonEdge :: Tgraph -> DEdge -> Tgraph -> Tgraph

relabelByCommonEdge g1 (x,y) g2 produces a relabeled version of g2 that is consistent with g1 on their overlap. The overlaping region must contain the common directed edge (x,y) without relabelling.

CAVEAT: The overlap must be a SINGLE tile-connected region in g1. (If the overlap contains more than one tile-connected region the result may not be a correct relabelling of g2)

relabelByEdges :: (Tgraph, DEdge) -> (Tgraph, DEdge) -> Tgraph

rrelabelByEdges (g1,e1) (g2,e2) produces a relabeled version of g2 that is consistent with g1 on their overlap. The overlaping region must contain the directed edge e1 in g1 and e2 in g2, which will be identified by the relabelling of g2.

CAVEAT: The overlap must be a SINGLE tile-connected region in g1. (If the overlap contains more than one tile-connected region the result may not be a correct relabelling of g2)

createVMapUsing :: (Tgraph, TileFace) -> (Tgraph, TileFace) -> Relabelling

createVMapUsing (g1,fc1) (g2,fc2) creates a relabelling map to make g2 consistent with g1 in a region of overlap if this is possible. fc1 and fc2 should be matching face types and have at least 2 vertices the same (without relabelling), with fc1 in g1 and fc2 in g2. g2 must have no vertices in common with g1 at the outset except for the 2 common vertices in fc1 and fc2 The resulting map when applied to fc2 will be identical to fc1. The common overlap needs to be a SINGLE region of tile-connected faces in g1. The common region must contain fc1 in g1 and fc2 in g2 The resulting map gives the required relabelling for g2 This will not produce correct results if there are other overlaps not tile-connected in g1 to the region containing fc1

addToVmap :: Tgraph -> [TileFace] -> [TileFace] -> [TileFace] -> Relabelling -> Relabelling

addToVmap is used by createVMapUsing to build a relabelling map. (addToVmap g processing tried awaiting vmap) produces a new vmap to be used for relabelling faces being matched against g. A match occurs if a relabeled face matches the face type and at least 2 corresponding vertices of a face in g. Argument g is the Tgraph being matched against. Argument processing is a list of faces to be matched next and each has an edge in common with at least one previously matched face. Argument tried is a list of faces that have been checked for a matching face in g but none found. They may still be on the boundary of g. Argument awaiting is a list of faces that have not yet been tried for a match and are not tile-connected to any faces already matched. Argument vmap is a vertex to vertex mapping used to store relabelling information during the process.

The idea is that from a single matched starting face we process faces that share an edge with a previously matched face. We prioritorize processing faces that have a match in g (with 2 matching vertices). If a face is tried but has no such match, it is stacked up in the tried list to await a check (with addToVmapBdCheck) for boundary cases after all faces in the overlap have been processed (when processing == []). If a processed face has an edge in common with a face in g but the faces do not match, this raises an error (indicating a mismatch on the overlap).

addToVmapBdCheck :: Boundary -> [TileFace] -> Relabelling -> Relabelling

addToVmapBdCheck is an auxiliary function for addToVmap to check boundary cases. When it is called, all faces in the overlap have been processed. The awaiting list is then dropped because: faces in awaiting cannot have an edge in common with the processed overlap and therefore cannot have a boundary edge. (If they have a single boundary vertex, this already has a relabelling in vMap) It remains to process the tried faces to see if they are on the boundary. If not, they can be ignored, but if they have a boundary edge, we need to find if the third vertex coincides with a neighbouring boundary vertex, in case this vertex also needs relabelling. All this assumes the overlap was a single region (satisfying Tgraph proerties)

checkForBoundaryThirdV :: Boundary -> TileFace -> Maybe (Vertex, Vertex)

checkForBoundaryThirdV bdry fc - fc must be a new face sharing a boundary edge in bdry. It checks if the third vertex of fc coincides with a third vertex on the neighbouring boundary (with possibly different label). If this is the case, it returns Just(v,v') to indicate the relabelling required for third vertex v (of fc to match with the found vertex v' on the boundary. Returns Nothing if such a vertex is not found. CAVEAT: No location checking is done to check for a non-local touching vertex.

noConflictCheck :: TileFace -> Boundary -> Bool

noConflictCheck fc bdry - f must be a new face sharing a boundary directed edge in bdry. It returns true if f has no conflicts with its neighbouring faces on the boundary.

thirdVAngles :: DEdge -> TileFace -> (Vertex, (Int, Int))

thirdVAngles e f - the directed edge e must be one of the directed edges of face f. The result is the third vertex of f paired with the two internal angles of the face on the directed edge (in order). The internal angles are expressed as a pair of integer multiples of ttangle (so 1,2,or 3). These numbers are used by findThirdV.

thirdVertexIn :: Tgraph -> TileFace -> Maybe [(Vertex, Vertex)]

thirdVertexIn g f tries to find a face in g corresponding to face f, where at least 2 vertices have to match. If successful, it returns Just [(v,v')] where v needs to be relabeled to v' for a 3 verrtex match, OR it may return Just [] if there is already a 3 vertex match. A return of Nothing indicates a failure to find a match.

pairing :: DEdge -> Tgraph -> Tgraph -> Maybe (TileFace, TileFace)

pairing e g1 g2 - finds a a face in g1 with directed edge and a face in g2 with directed edge e It returns an empty list if the faces do not match,

hasEdge :: DEdge -> TileFace -> Bool

hasEdge e f returns True if directed edge e is one of the directed edges of face f

initVMap :: TileFace -> TileFace -> Relabelling

initVMap f1 f2 - creates a relabelling (vertex to vertex mapping) so that if applied to face f2, the vertices will match with face f1 exactly.

matchFace :: TileFace -> [TileFace] -> Maybe TileFace

matchFace f1 fcs - looks for a face in fcs that has a directed edge in common with f1. If such a face is found it checks that it matches (except possibly for one vertex label) Returns Just f if f is the matched face, but an error f does not match. Retruns Nothing if no face in fcs has a directed edge in common with f1

hasEdgeIn :: [DEdge] -> TileFace -> Bool

hasEdgeIn es fc - is True if fc has a directed edge in the list of edges es.

twoVMatch :: TileFace -> TileFace -> Bool

twoVMatch f1 f2 is True if the two tilefaces are the same except for a single vertex label possibly not matching.

sameFaceType :: HalfTile rep1 -> HalfTile rep2 -> Bool

sameFaceType f1 f2 is True if f1 and f2 are of the same form (i.e. both LD or both RD or both LK or both RK)

module Tgraph.Prelude

module Tgraph.Decompose

module Tgraph.Compose

module Tgraph.Force

module Tgraph.Convert