TileFigExamples

Description

Documentation

label :: (Renderable (Text n) b, RealFloat n, Typeable n) => String -> Point V2 n -> QDiagram b V2 n Any

Used for adding text at a point

Figures for pieces

thePieces :: [Piece]

a list of the four Pieces

piecesFig :: Diagram B

drawn edges of 4 pieces in a row

piecesFig2 :: QDiagram B V2 Double Any

filled 4 pieces in a row

markedTiles :: Diagram B

figure showing origins and markings on tiles

markedTiles2 :: Diagram B

another figure showing origins and markings on tiles

newPiecesFig :: Diagram B

figure showing the four labelled pieces each with join edge and origin shown

tileIntro :: Diagram B

diagram combining markedTiles2 and newPiecesFig

Figures for decompositions and inflations

fourDecomps :: Diagram B

figure showing 4 decompositions in a column for each of the four pieces

fiveInflate :: Diagram B

example of inflate in action with 5 chosen inflation steps from a left dart. This shows inital and final piece together on the left, and 5 decomposition of the final piece on the right.

get5inflate4 :: Located Piece -> Diagram B

show just the first 5 alternatives of 4-fold inflations of a located piece (previously called inflatefig)

fiveAlternatives :: Diagram B

diagram showing first five alternatives of 4-fold inflations of a right dart

suns :: [Patch]

An infinite list of patches of increasingly decomposed sun

sun5 :: Patch

a patch of a 5 times decomposed sun

sun6 :: Patch

a patch of a 6 times decomposed sun

sun6Fig :: Diagram B

diagram for sun6

sun5Over6Fig :: Diagram B

diagram overlaying sun5 in red atop sun6

experimentFig :: Diagram B

Using experiment (defined in Tilelib) on sun6 clearly illustrates the embedded sun5

twoLevelsFig :: Diagram B

Using experiment (defined in Tilelib) on sun4 clearly illustrates the embedded sun3

dartWingFig :: Diagram B

figure showing two types of dart wing vertices (largeKiteCentre, largeDartBase)

Colour-filled examples

filledSun6 :: Diagram B

using fillDK'

newFillSun6 :: Diagram B

using fillDK

threeColouredShapes :: [Diagram B]

list of 3 diagrams for colour filled star,sun kite respectivly

exampleTriple1 :: Diagram B

diagram of three coloured shapes in a row

exampleTriple2 :: Diagram B

diagram of three coloured shapes in an arc

Swatches and Samples for colour-filled patches

type Sample = (Colour Double, Colour Double, Colour Double)

A sample abbreviates triples of colours (used for Dart,Kite,Grout (Edges) respectively

drawSample :: Sample -> Diagram B

The sample is used to fill a 6 times decomposed sun

sampleFig0 :: Diagram B

sampleFig0 is a figure drawn with drawSample

drawSampleSmall :: Sample -> Diagram B

The sample is used to fill a 7 times decomposed sun (so smaller kites and darts than drawSample).

drawSampleTiny :: Sample -> Diagram B

The sample is used to fill an 8 times decomposed sun (so muuch smaller kites and darts than drawSample).

centreCropA4 :: Diagram B -> Diagram B

crop a diagram to half width and centred with A4 portrait dimensions. Ideal for decomposed suns.

a4Small :: Sample -> Diagram B

Make an A4 diagram using sample colours with drawSampleSmall

a4Tiny :: Sample -> Diagram B

Make an A4 diagram using sample colours with drawSampleTiny

blackAndWhite :: Diagram B

diagram for black and white example using a4Tiny

migA4Tiny :: Diagram B

diagram for darkmagenta, indigo, gold example using a4Tiny

bryA4Small :: Diagram B

diagram for darkblue,red ,yellow example using a4Small

type Swatch = [Sample]

A swatch is a list of samples which are used to fill sun5s by drawSwatch

drawSwatch :: Int -> Swatch -> Diagram B

drawSwatch n sw produces a sun5 filled example for each sample in the swatch sw. These are combined into a single diagram in rows of length n

swatch0 :: Swatch

an example swatch

swatchFig0 :: Diagram B

figure for example swatch