TileLib

Description

This module includes the basic operations for creating and drawing finite patches of Penrose's Dart and Kite tilings. It includes a decompose operation, drawPatch, fillDK and colourDKG, sun and star example patches.

Documentation

type Piece = HalfTile (V2 Double)

Piece type for tile halves: Left Dart, Right Dart, Left Kite, Right Kite with a vector from their origin along the join edge where origin for a dart is the tip, origin for a kite is the acute angle tip. Using Imported polymorphic HalfTile

getJVec :: Piece -> V2 Double

get the vector representing the join edge in the direction away from the origin of a piece

ldart :: Piece

ldart,rdart,lkite,rkite are the 4 pieces (oriented along the x axis).

rdart :: Piece

ldart,rdart,lkite,rkite are the 4 pieces (oriented along the x axis).

lkite :: Piece

ldart,rdart,lkite,rkite are the 4 pieces (oriented along the x axis).

rkite :: Piece

ldart,rdart,lkite,rkite are the 4 pieces (oriented along the x axis).

phi :: Double

All edges are powers of the golden section phi. We also have the interesting property of the golden section that phi^2 == phi + 1 and so 1/phi = phi-1 (also phi^3 = 2phi +1 and 1/phi^2 = 2-phi)

ttangle :: Int -> Angle Double

All angles used are multiples of tt where tt is a tenth of a turn (so tau divided by 10 rad or 36 deg). ttangle n is n multiples of tt. Angles are from positive x axis anticlockwise

pieceEdges :: Piece -> [V2 Double]

produces a list of the two non-join tile edges of a piece. Perhaps confusingly we regard left and right of a dart differently from left and right of a kite. This is in line with common sense view but darts are reversed from origin point of view. Going clockwise round the origin the Right Dart comes before Left Dart, but the Left Kite comes before Right Kite. This is manifest in only pieceEdges, wholeTileEdges, compose and decompPiece

wholeTileEdges :: Piece -> [V2 Double]

the 4 tile edges of a completed half-tile piece (used for colour fill). These are clockwise from the origin of the tile.

drawPiece :: Piece -> Diagram B

drawing lines for the 2 non-join edges of a piece

fillDK' :: Colour Double -> Colour Double -> Piece -> Diagram B

fillDK' dcol kcol pc fills the whole tile when pc is a left half-tile, darts are filled with colour dcol and kites with colour kcol. (Right half-tiles produce nothing, so whole tiles are not drawn twice)

drawJPiece :: Piece -> Diagram B

same as drawPiece but with added join edge (also fillable as a loop)

fillPiece :: Colour Double -> Piece -> Diagram B

fillPiece col piece - fills piece with colour col without drawing any lines

fillDK :: Colour Double -> Colour Double -> Piece -> Diagram B

fillDK dcol kcol piece - draws and fills the half-tile piece with colour dcol for darts and kcol for kites. Note the order D K.

dashJPiece :: Piece -> Diagram B

same as drawPiece but with join edge added as dashed-line

drawJ :: Piece -> Diagram B

draw join edge only

experiment :: Piece -> Diagram B

experiment uses a different rule for drawing half tiles. This clearly displays the larger kites and darts. Half tiles are first drawn with dashed lines, then certain edges are overlayed to emphasise them. Half darts have the join edge emphasised in red, while Half kites have the long edge emphasised in black.

type Patch = [Located Piece]

A patch is a list of Located pieces (the point associated with each piece locates its originV)

patchWith :: (Piece -> Diagram B) -> Patch -> Diagram B

turn a patch into a diagram using the first argument for drawing pieces

drawPatch :: Patch -> Diagram B

special case of patchWith - turn patches to diagrams with drawPiece

dashJPatch :: Patch -> Diagram B

special case of patchWith - turn patches to diagrams with dashJPiece

colourDKG :: (Colour Double, Colour Double, Colour Double) -> Patch -> Diagram B

colourDKG (c1,c2,c3) p fill in a patch p with colour c1 for darts, colour c2 for kites and colour c3 for grout (that is, the non-join edges). Note the order D K G.

decompose :: Patch -> Patch

Decomposing splits each located piece in a patch into a list of smaller located pieces to create a refined patch Decomposition is uniquely determined.

decompPiece :: Located (HalfTile (V2 Double)) -> [Located (HalfTile (V2 Double))]

Decomposing a located piece

decompositions :: Patch -> [Patch]

Create an infinite list of increasing decompositions of a patch

compChoices :: Located Piece -> [Located Piece]

compChoices applied to a single located piece produces a list of alternative located pieces NOT a Patch. Each of these is a larger scale single piece with a location such that when decomposed the original piece in its original position is part of the decomposition)

compNChoices :: Int -> Located Piece -> [Located Piece]

compNChoices n lp - gives a list of all the alternatives after n compChoices starting with lp

inflate :: Located Piece -> [Located Piece]

inflate is a deprecated name for what is now compChoices

inflations :: Int -> Located Piece -> [Located Piece]

a deprecated name for what is now compNChoices

penta :: Patch -> Patch

combine 5 copies of a patch (each rotated by ttangle 2 successively) (ttAngle 2 is 72 degrees) Must be used with care to avoid creating a nonsense patch

sun :: Patch

sun is a patch with five kites sharing common origin (base of kite)

star :: Patch

star is a patch with five darts sharing common origin (tip of dart)

rotations :: (Transformable a, V a ~ V2, N a ~ Double) => [Int] -> [a] -> [a]

rotations takes a list of integers (ttangles) for respective rotations of items in the second list (things to be rotated). This includes Diagrams, Patches, VPatches The integer list can be shorter than the list of items - the remaining items are left unrotated.

scales :: (Transformable a, V a ~ V2, N a ~ Double) => [Double] -> [a] -> [a]

scales takes a list of doubles for respective scalings of items in the second list (things to be scaled). This includes Diagrams, Patches, VPatches The list of doubles can be shorter than the list of items - the remaining items are left unscaled.

phiScales :: (Transformable a, V a ~ V2, N a ~ Double) => [a] -> [a]

increasing scales by phi along a list starting with 1

phiScaling :: (Transformable a, V a ~ V2, N a ~ Double) => Double -> [a] -> [a]

increasing scales by phi along a list starting with given first argument

Orphan instances

Transformable a => Transformable (HalfTile a)
Making HalfTiles (and therefore Pieces and Patches) transformable

Instance details

Methods

transform :: Transformation (V (HalfTile a)) (N (HalfTile a)) -> HalfTile a -> HalfTile a