Tgraph.Prelude

Description

Introduces Tgraphs and includes operations on vertices, edges and faces as well as Tgraphs. Also includes functions to calculate (relative) locations of vertices (createVPoints) and re-exports module HalfTile

Documentation

type Mapping = Map

Mapping abbreviates Data.Map.Map

Tgraphs

type Vertex = Int

Tgraph vertices

type DEdge = (Vertex, Vertex)

directed edge

type TileFace = HalfTile (Vertex, Vertex, Vertex)

Tgraph faces (vertices clockwise starting with tile origin vertex) a specialisation of HalfTile

data Tgraph

A Tgraph contains vertices, and faces (each are lists treated as sets with no repetitions). Every vertex must be a face vertex and vice versa

Constructors

Tgraph

Fields

vertices :: [Vertex]

faces :: [TileFace]

makeUncheckedTgraph :: [TileFace] -> Tgraph

Creates a (possibly invalid) Tgraph from a list of faces by calculating vertices. It does not perform checks on the faces. Use checkedTgraph to perform checks.

checkedTgraph :: [TileFace] -> Tgraph

Creates a Tgraph from a list of faces AND checks for edge conflicts and crossing boundaries and connectedness with checkTgraphProps. (No crossing boundaries and connected implies tile-connected). Produces an error if a check fails.

checkTgraphProps :: [TileFace] -> Either [String] Tgraph

Checks a list of faces for edge loops, edge conflicts and crossing boundaries and connectedness. (No crossing boundaries and connected implies tile-connected) Returns Right g where g is a Tgraph on passing checks. Returns Left lines if a test fails, where lines describes the problem found.

selectFaces :: [TileFace] -> Tgraph -> Tgraph

selects faces from a Tgraph (removing any not in the list), but checks resulting Tgraph for required properties e.g. connectedness and no crossing boundaries.

removeFaces :: [TileFace] -> Tgraph -> Tgraph

removes faces from a Tgraph, but checks resulting Tgraph for required properties e.g. connectedness and no crossing boundaries.

removeVertices :: [Vertex] -> Tgraph -> Tgraph

removeVertices vs g - removes any vertex in the list vs from g by removing all faces at those vertices. Resulting Tgrpah is checked for required properties e.g. connectedness and no crossing boundaries.

selectVertices :: [Vertex] -> Tgraph -> Tgraph

selectVertices vs g - removes any face that does not have a vertex in the list vs from g. Resulting Tgrpah is checked for required properties e.g. connectedness and no crossing boundaries.

nullGraph :: Tgraph -> Bool

is the graph empty?

Tests and Tgraph properties

findEdgeLoops :: [TileFace] -> [Vertex]

Returns any repeated vertices in a single tileface for a list of tilefaces.

hasEdgeLoops :: [TileFace] -> Bool

Checks if there are repeated vertices within a tileface for a list of tilefaces. Returns True if there are any.

duplicates :: Eq a => [a] -> [a]

duplicates finds duplicated items in a list

conflictingDedges :: Tgraph -> [DEdge]

conflictingDedges g returns a list of conflicting directed edges of the faces in g (which should be null)

conflictingLengthEdges :: Tgraph -> [DEdge]

conflictingLengthEdges g returns a list of conflicting lengthed edges of the faces in g (which should be null)

edgeConflicts :: Tgraph -> Bool

predicate - true if there are edge conflicts in a Tgraph

crossingBVs :: Tgraph -> [Vertex]

crossingBVs g returns a list of vertices with crossing boundaries (which should be null).

crossingBoundaries :: Tgraph -> Bool

There are crossing boundaries if vertices occur more than once at the start of all boundary directed edges (or more than once at the end of all boundary directed edges).

connected :: Tgraph -> Bool

Predicate to check a Tgraph is a connected graph.

connectedBy :: Eq a => [(a, a)] -> a -> [a] -> [a]

Auxiliary function for calculating connectedness by depth first search. connectedBy edges v verts returns the sublist of verts connected to v by a chain of edges

Face and Vertex Operations

ldarts :: Tgraph -> [TileFace]

selecting left darts, right darts, left kite, right kites from a Tgraph

rdarts :: Tgraph -> [TileFace]

selecting left darts, right darts, left kite, right kites from a Tgraph

lkites :: Tgraph -> [TileFace]

selecting left darts, right darts, left kite, right kites from a Tgraph

rkites :: Tgraph -> [TileFace]

selecting left darts, right darts, left kite, right kites from a Tgraph

valencyD :: Tgraph -> Vertex -> Int

directed edge valency of a vertex

faceVs :: TileFace -> (Vertex, Vertex, Vertex)

triple of face vertices in order clockwise - tileRep specialised to TileFace

faceVList :: TileFace -> [Vertex]

list of (three) face vertices in order clockwise

firstV :: TileFace -> Vertex

Whilst first, second and third vertex of a face are obvious (clockwise), it is often more convenient to refer to the originV (=firstV), oppV (the vertex at the other end of the join edge), and wingV (the remaining vertex not on the join edge)

secondV :: TileFace -> Vertex

Whilst first, second and third vertex of a face are obvious (clockwise), it is often more convenient to refer to the originV (=firstV), oppV (the vertex at the other end of the join edge), and wingV (the remaining vertex not on the join edge)

thirdV :: TileFace -> Vertex

Whilst first, second and third vertex of a face are obvious (clockwise), it is often more convenient to refer to the originV (=firstV), oppV (the vertex at the other end of the join edge), and wingV (the remaining vertex not on the join edge)

originV :: TileFace -> Vertex

Whilst first, second and third vertex of a face are obvious (clockwise), it is often more convenient to refer to the originV (=firstV), oppV (the vertex at the other end of the join edge), and wingV (the remaining vertex not on the join edge)

wingV :: TileFace -> Vertex

Whilst first, second and third vertex of a face are obvious (clockwise), it is often more convenient to refer to the originV (=firstV), oppV (the vertex at the other end of the join edge), and wingV (the remaining vertex not on the join edge)

oppV :: TileFace -> Vertex

Whilst first, second and third vertex of a face are obvious (clockwise), it is often more convenient to refer to the originV (=firstV), oppV (the vertex at the other end of the join edge), and wingV (the remaining vertex not on the join edge)

indexV :: Int -> TileFace -> Int

indexV finds the index of a vertex in a face (firstV -> 0, secondV -> 1, thirdV -> 2)

nextV :: Vertex -> TileFace -> Vertex

nextV returns the next vertex in a face going clockwise from v where v must be a vertex of the face

prevV :: Vertex -> TileFace -> Vertex

prevV returns the previous vertex in a face (i.e. next going anti-clockwise) from v where v must be a vertex of the face

isAtV :: Vertex -> TileFace -> Bool

isAtV v fc asks if a face fc has v as a vertex

hasVIn :: [Vertex] -> TileFace -> Bool

hasVIn vs fc - asks if face fc has an element of vs as a vertex

makeNewVs :: Int -> [Vertex] -> [Vertex]

given existing vertices vs, create n new vertices

makeNewV :: [Vertex] -> Vertex

return one new vertex

Edge Operations

faceDedges :: TileFace -> [DEdge]

directed edges (clockwise) round a face

reverseD :: DEdge -> DEdge

opposite directed edge

firstE :: TileFace -> DEdge

Whilst first, second and third edges are obvious (always clockwise), it is often more convenient to refer to the joinE (join edge), shortE (the short edge which is not a join edge), and longE (the long edge which is not a join edge). These are also directed clockwise. joinOfTile also returns the join edge but in the direction away from the origin

secondE :: TileFace -> DEdge

Whilst first, second and third edges are obvious (always clockwise), it is often more convenient to refer to the joinE (join edge), shortE (the short edge which is not a join edge), and longE (the long edge which is not a join edge). These are also directed clockwise. joinOfTile also returns the join edge but in the direction away from the origin

thirdE :: TileFace -> DEdge

Whilst first, second and third edges are obvious (always clockwise), it is often more convenient to refer to the joinE (join edge), shortE (the short edge which is not a join edge), and longE (the long edge which is not a join edge). These are also directed clockwise. joinOfTile also returns the join edge but in the direction away from the origin

joinE :: TileFace -> DEdge

the join edge of a face in the clockwise direction going round the face (see also joinOfTile).

Whilst first, second and third edges are obvious (always clockwise), it is often more convenient to refer to the joinE (join edge), shortE (the short edge which is not a join edge), and longE (the long edge which is not a join edge). These are also directed clockwise. joinOfTile also returns the join edge but in the direction away from the origin

shortE :: TileFace -> DEdge

The short edge of a face in the clockwise direction going round the face. This is the non-join short edge for darts.

Whilst first, second and third edges are obvious (always clockwise), it is often more convenient to refer to the joinE (join edge), shortE (the short edge which is not a join edge), and longE (the long edge which is not a join edge). These are also directed clockwise. joinOfTile also returns the join edge but in the direction away from the origin

longE :: TileFace -> DEdge

The long edge of a face in the clockwise direction going round the face. This is the non-join long edge for kites.

Whilst first, second and third edges are obvious (always clockwise), it is often more convenient to refer to the joinE (join edge), shortE (the short edge which is not a join edge), and longE (the long edge which is not a join edge). These are also directed clockwise. joinOfTile also returns the join edge but in the direction away from the origin

joinOfTile :: TileFace -> DEdge

The join edge of a face but directed from the origin (not clockwise for RD and LK)

Whilst first, second and third edges are obvious (always clockwise), it is often more convenient to refer to the joinE (join edge), shortE (the short edge which is not a join edge), and longE (the long edge which is not a join edge). These are also directed clockwise. joinOfTile also returns the join edge but in the direction away from the origin

facePhiEdges :: TileFace -> [DEdge]

The phi edges of a face (both directions) which is long edges for darts, and join and long edges for kites

faceNonPhiEdges :: TileFace -> [DEdge]

The non-phi edges of a face (both directions) which is short edges for kites, and join and short edges for darts

matchingE :: (TileFace -> DEdge) -> TileFace -> TileFace -> Bool

matchingE etype fc is a predicate on tile faces where etype finds a particular edge type of a face (etype could be joinE or longE or shortE for example). This is True for fc' if fc' has an etype edge matching the (reversed) etype edge of fc

matchingLongE :: TileFace -> TileFace -> Bool

special cases of matchingE etype where etype is longE, shortE, and joinE

matchingShortE :: TileFace -> TileFace -> Bool

special cases of matchingE etype where etype is longE, shortE, and joinE

matchingJoinE :: TileFace -> TileFace -> Bool

special cases of matchingE etype where etype is longE, shortE, and joinE

hasDEdge :: DEdge -> TileFace -> Bool

hasDEdge e f returns True if directed edge e is one of the directed edges of face f

hasDEdgeIn :: [DEdge] -> TileFace -> Bool

hasDEdgeIn es fc - is True if fc has a directed edge in the list of edges es.

graphDedges :: Tgraph -> [(Vertex, Vertex)]

A list of all the directed edges of a graph (going clockwise round faces)

phiEdges :: Tgraph -> [(Vertex, Vertex)]

phiEdges returns a list of the phi-edges of a Tgraph (= long non-join edges of the faces). This includes both directions of each edge.

nonPhiEdges :: Tgraph -> [(Vertex, Vertex)]

nonPhiEdges returns a list of the shorter non-join edges of a Tgraph. This includes both directions of each edge.

graphEdges :: Tgraph -> [(Vertex, Vertex)]

graphEdges returns a list of all the edges of a Tgraph (both directions of each edge).

bothDir :: [DEdge] -> [DEdge]

bothDir adds missing reverse directed edges to a list of directed edges and then removes duplicates

bothDir' :: [DEdge] -> [DEdge]

bothDir' adds the reverse directed edges to a list of directed edges without checking for duplicates

boundaryDedges :: Tgraph -> [(Vertex, Vertex)]

boundaryDedges g are missing reverse directed edges in graphDedges g (these are single directions only) Direction is such that a face is on LHS and exterior is on RHS of each boundary directed edge.

boundaryEdges :: Tgraph -> [(Vertex, Vertex)]

boundary edges are face edges not shared by 2 faces (but both directions).

internalEdges :: Tgraph -> [(Vertex, Vertex)]

internal edges are shared by two faces = all edges except boundary edges

edgeNb :: TileFace -> TileFace -> Bool

two tile faces are edge neighbours

Vertex Location Calculation

createVPoints :: [TileFace] -> Mapping Vertex (Point V2 Double)

createVPoints: processes a list of faces to associate points for each vertex. Faces must be tile-connected. It aligns the join of the first face on the x-axis. Returns a vertex-to-point Map.

addVPoints :: [TileFace] -> [TileFace] -> Mapping Vertex (Point V2 Double) -> Mapping Vertex (Point V2 Double)

addVPoints readyfaces fcOther vpMap. The first argument list of faces (readyfaces) contains the ones being processed next in order where each will have at least two known vertex locations. The second argument list of faces (fcOther) are faces that have not yet been added and may not yet have known vertex locations. The third argument is the mapping of vertices to points. This is used in tryUpdate as well as createVPoints.

initJoin :: TileFace -> Mapping Vertex (Point V2 Double)

initJoin fc initialises a vertex to point mapping with locations for the join edge vertices of fc with originV fc at the origin and aligned along the x axis. (Used to initialise createVPoints)

find3Locs :: (Vertex, Vertex, Vertex) -> Mapping Vertex (Point V2 Double) -> (Maybe (Point V2 Double), Maybe (Point V2 Double), Maybe (Point V2 Double))

lookup 3 vertex locations in a vertex to point map.

thirdVertexLoc :: TileFace -> Mapping Vertex (Point V2 Double) -> Maybe (Vertex, Point V2 Double)

New Version - Assumes all edge lengths are 1 or phi. It now uses signorm to produce vectors of length 1 rather than rely on relative lengths. (Requires ttangle and phi from TileLib).

thirdVertexLoc fc vpMap where fc is a tileface and vpMap associates points with vertices (positions) It looks up all 3 vertices of fc in vpMap hoping to find at least 2 of them, it then returns Just pr where pr is an association pair for the third vertex. If all 3 are found, returns Nothing. If none or one found this is an error (a non tile-connected face)

SubTgraphs

data SubTgraph

SubTgraph - introduced to allow tracking of subsets of faces in both force and decompose oerations. A SubTgraph has a main Tgraph (fullgraph) and a list of subsets of faces (trackedSubsets). The list allows for tracking different subsets of faces at the same time

Constructors

SubTgraph

Fields

fullGraph :: Tgraph

trackedSubsets :: [[TileFace]]

makeSubTgraph :: Tgraph -> [[TileFace]] -> SubTgraph

makeSubTgraph g trackedlist creates a SubTgraph from a Tgraph g from trackedlist where each list in trackedlist is a subset of the faces of g. Any faces not in g are ignored.