Tgraph.Relabelling

Description

This module includes relabelling functions for Tgraphs and a guided union of Tgraphs

Documentation

type Relabelling = Map Vertex Vertex

Relabelling is a special case of mappings from vertices to vertices. We use the identity map for vertices not found in the mapping domain (see relabelV). Relabellings are expected to be 1-1 on their domain.

relabelGraph :: Relabelling -> Tgraph -> Tgraph

Uses a relabelling to change vertices in a Tgraph. relabelGraph vmap g will produce a valid Tgraph provided: g is a valid Tgraph, and the mapping vmap extended with the identity is 1-1 on vertices in g. (Vertices of g that are not in the domain of the mapping are left unchanged.)

relabelFace :: Relabelling -> TileFace -> TileFace

Uses a relabelling to relabel the three vertices of a face. Any vertex not in the domain of the mapping is left unchanged. The mapping should be 1-1 on the 3 vertices to avoid creating a self loop edge.

relabelV :: Relabelling -> Vertex -> Vertex

relabelV vmap v. Uses relabelling vmap to find a replacement for v (leaves as v if none found). I.e relabelV turns a relabelling into a total function using identity for undefined cases.

all3 :: (a -> b) -> (a, a, a) -> (b, b, b)

Applies a function to all three elements of a triple.

partRelabelAvoid :: [Vertex] -> Tgraph -> Tgraph

partRelabelAvoid vs g - produces a new Tgraph from g by relabelling. Any vertex in g that is in the list vs will be changed to a new vertex that is neither in g nor in the list vs. Vertices in g that are not in vs will remain the same.

relabelAny :: Tgraph -> Tgraph

Relabel all vertices in a Tgraph using new labels 1..n (where n is the number of vertices).

tryUnionGraphs :: (Tgraph, DEdge) -> (Tgraph, DEdge) -> ReportFail Tgraph

tryUnionGraphs (g1,e1) (g2,e2) - where edge e1 is in g1 and e2 is in g2, checks if g2 can be relabelled to produce a common single region of overlap with g1 (with e2 relabelled to e1). If so then the result is Right g where g is the union of the faces. Otherwise the result is Left lines where lines explains the problem.

unionGraphs :: (Tgraph, DEdge) -> (Tgraph, DEdge) -> Tgraph

unionGraphs (g1,e1) (g2,e2) - where edge e1 is in g1 and e2 is in g2, checks if g2 can be relabelled to produce a common single region of overlap with g1 (with e2 relabelled to e1). If so then the result is Right g where g is the union of the faces. Otherwise an error is raised.

matchByCommonEdge :: Tgraph -> DEdge -> Tgraph -> Tgraph

matchByCommonEdge g1 e g2 produces a relabelled version of g2 that is consistent with g1 on their overlap. The overlaping region must contain the common directed edge e without relabelling. This produces an error if a mismatcch is found in the overlap.

CAVEAT: The overlap must be a SINGLE tile-connected region in g1. (If the overlap contains more than one tile-connected region the result may not be a correct relabelling of g2)

tryMatchByCommonEdge :: Tgraph -> DEdge -> Tgraph -> ReportFail Tgraph

tryMatchByCommonEdge g1 e g2 produces either Right g where g is a relabelled version of g2 that is consistent with g1 on their overlap or Left lines if there is a mismatch (lines explaining the problem). The overlaping region must contain the common directed edge e without relabelling.

CAVEAT: The overlap must be a SINGLE tile-connected region in g1. (If the overlap contains more than one tile-connected region the result may not be a correct relabelling of g2)

matchByEdges :: (Tgraph, DEdge) -> (Tgraph, DEdge) -> Tgraph

matchByEdges (g1,e1) (g2,e2) produces a relabelled version of g2 that is consistent with g1 on their overlap. The overlapping region must contain the directed edge e1 in g1. The edge e2 in g2 will be identified with e1 by the relabelling of g2. This produces an error if a mismatcch is found in the overlap.

CAVEAT: The overlap must be a SINGLE tile-connected region in g1. (If the overlap contains more than one tile-connected region the result may not be a correct relabelling of g2)

tryMatchByEdges :: (Tgraph, DEdge) -> (Tgraph, DEdge) -> ReportFail Tgraph

tryMatchByEdges (g1,e1) (g2,e2) produces either Right g where g is a relabelled version of g2 that is consistent with g1 on their overlap or Left lines if there is a mismatch (lines explaining the problem). The overlapping region must contain the directed edge e1 in g1. The edge e2 in g2 will be identified with e1 by the relabelling of g2.

CAVEAT: The overlap must be a SINGLE tile-connected region in g1. (If the overlap contains more than one tile-connected region the result may not be a correct relabelling of g2)

findRelabelling :: (Tgraph, TileFace) -> (Tgraph, TileFace) -> ReportFail Relabelling

findRelabelling is an auxiliary function for tryMatchByEdges and tryMatchByCommonEdge. findRelabelling (g1,fc1) (g2,fc2) - fc1 and fc2 should be matching face types, with fc1 in g1 and fc2 in g2. g2 must have no vertices in common with g1 except for (possibly) vertices in fc2. The result is either Right rel where rel is a relabelling map to make g2 consistent with g1 in a region of overlap if this is possible, or Left lines if there is a mismatch (lines explaining the problem). In the successful case rel when applied to fc2 will be identical to fc1.

CAVEAT: The common overlap needs to be a SINGLE region of tile-connected faces in g1. The common region must contain fc1 in g1 and fc2 in g2. This may not produce correct results if there are other overlaps not tile-connected in g1 to the region containing fc1

initRelabelling :: TileFace -> TileFace -> Relabelling

initRelabelling f1 f2 - creates a relabelling so that if applied to face f2, the vertices will match with face f1 exactly. It does not check that the tile faces have the same form (LK,RK,LD,RD).

differences :: [(Vertex, Vertex)] -> [(Vertex, Vertex)]

selects only non-matching pairs from a list

addRelabel :: Tgraph -> [TileFace] -> [TileFace] -> [TileFace] -> Relabelling -> ReportFail Relabelling

addRelabel is used by findRelabelling to build a relabelling map which can fail, producing Left lines. In the successful case (addRelabel g processing tried awaiting vmap) produces a Right rel where rel is the required relabelling. The arguments are: g - the Tgraph being matched against. processing - a list of faces to be matched next where each has an edge in common with at least one previously matched face (or it is the starting face). tried - a list of faces that have been checked for a matching face in g but none found. They will each have an edge in the boundary of g. awaiting - a list of faces that have not yet been tried for a match and are not tile-connected to any faces already matched. vmap - a vertex to vertex mapping used to store relabelling information during the process.

The idea is that from a single matched starting face we process faces that share an edge with a previously matched face. We prioritorize processing faces that have a match in g (with 2 matching vertices). If a face is tried but has no such match, it is stacked up in the tried list to await a check (with addRelabelBdCheck) boundary cases after all faces in the overlap have been processed (i.e. when processing == []). If a processed face has an edge in common with a face in g it has to match exactly apart from (possibly) the third vertex label, otherwise the faces do not match and this indicates a mismatch on the overlap and Left ... is returned.

addRelabelBdCheck :: Boundary -> [TileFace] -> Relabelling -> ReportFail Relabelling

addRelabelBdCheck is an auxiliary function for addRelabel to check boundary cases. When it is called, all faces in the overlap have been processed. The awaiting list is then dropped because: faces in awaiting cannot have an edge in common with the processed overlap and therefore cannot have a boundary edge. (If they have a single boundary vertex, this will get its relabelling from a neighbouring processed face) It remains to process the tried faces which must each have a boundary edge. We need to find if the third vertex (not on the boundary edge) coincides with a neighbouring boundary vertex, in case this vertex also needs relabelling. Such a face then has two (or possibly 3) boundary edges and is checked for conflicts with its neighbouring faces on the boundary.

CAVEAT: All this assumes the overlap is a single region.

checkForBoundaryThirdV :: Boundary -> TileFace -> ReportFail [(Vertex, Vertex)]

checkForBoundaryThirdV bdry fc - fc must be a new face sharing a boundary edge in bdry. It checks if the third vertex of fc coincides with a third vertex on the neighbouring boundary (with possibly different label). If this is NOT the case Right [] is returned. If this is the case, a check is made that the face has no conflicts with faces on the boundary. If there are no conflicts, Right [(v,v')] is returned to indicate the relabelling required for third vertex v (of fc) to match with the found vertex v' on the boundary. If there is a conflict, Left lines is returned (where lines explains the conflict)

CAVEAT: No location checking is done to check for a non-local touching vertex.

noConflictCheck :: TileFace -> Boundary -> Bool

noConflictCheck fc bdry - f must be a new face sharing a boundary directed edge in bdry. It returns true if f has no conflicts with its neighbouring faces on the boundary.

thirdVAngles :: DEdge -> TileFace -> (Vertex, (Int, Int))

thirdVAngles e f - the directed edge e must be one of the directed edges of face f. The result is the third vertex of f paired with the two internal angles of the face on the directed edge (in order). The internal angles are expressed as a pair of integer multiples of ttangle (so 1,2,or 3). These numbers are used by findThirdV.

matchFaceIn :: Tgraph -> TileFace -> ReportFail (Maybe TileFace)

matchFaceIn g f - looks for a face in g that corresponds to f (sharing a directed edge), If the corresponding face does not match properly (with twoVMatch) this stops the matching process returning Left ... to indicate a failed match. Otherwise it returns either Right (Just f) where f is the matched face or Right Nothing if there is no corresponding face.

twoVMatch :: TileFace -> TileFace -> Bool

twoVMatch f1 f2 is True if the two tilefaces are the same except for a single vertex label possibly not matching.

sameFaceType :: HalfTile rep1 -> HalfTile rep2 -> Bool

sameFaceType f1 f2 is True if f1 and f2 are of the same form (i.e. both LD or both RD or both LK or both RK)