Tgraphs

Description

This is the main module for Tgraphs which collects and exports the various other Tgraph modules and includes a definition of emplace and other experimental combinations.

Documentation

emplace :: Tgraph -> Tgraph

emplace does maximal composing with force and composeG, then applies decomposeG and force repeatedly back to the starting level. It produces the emplacement of influence of the argument graph.

emplacements :: Tgraph -> [Tgraph]

emplacements is best supplied with a maximally composed or near maximally composed graph It produces an infinite list of emplacements of the starting graph and its decompositions.

emplaceChoices :: Tgraph -> [Tgraph]

a version of emplace using makeChoices at the top level. after makeChoices we use emplace to attempt further compositions but with no further choices

makeChoices :: Tgraph -> [Tgraph]

makeChoices is a temporary tool which does not attempt to analyse choices for correctness. It can thus create some choices which will be incorrect. The unknowns returned from classifyDartWings can become largeKiteCentres or largeDartBases. This produces 2^n choices where n is the number of unknowns (excluding lone dart wing tips with valencyD 2).

compForce :: Tgraph -> Tgraph

compForce does a force then composeG but it by-passes the check on the composed graph because the argument was forced

allCompFs :: Tgraph -> [Tgraph]

allCompFs g produces a list of all forced compositions starting from g up to but excluding the empty graph

allComps :: Tgraph -> [Tgraph]

allComps g produces a list of all compositions starting from g up to but excluding the empty graph. This is not safe in general as it can fail by producing a non-connected graph or graph with crossing boundaries.

maxCompose :: Tgraph -> Tgraph

maxCompose and maxFCompose produce a maximal graph.

maxFCompose :: Tgraph -> Tgraph

maxCompose and maxFCompose produce a maximal graph.

removeIncompleteTiles :: Tgraph -> Tgraph

remove halftile faces that do not have their matching half tile

relabelGraph :: Mapping Vertex Vertex -> Tgraph -> Tgraph

Uses a vertex to vertex mapping to change vertices in a Tgraph relabelGraph vmap g will produce a valid Tgraph provided: g is a valid Tgraph, and the mapping vmap is 1-1 with no output vertices in common with the vertices of g. Vertices of g that are not in the domain of the mapping are left unchanged.

relabelFace :: Mapping Vertex Vertex -> TileFace -> TileFace

Uses a vertex to vertex mapping to relabelxf the three vertices of a face Any vertex not in the domain of the mapping is left unchanged. The mapping should be 1-1 to avoid creating a self loop edge.

relabelV :: Mapping Vertex Vertex -> Vertex -> Vertex

relabelV vmap v. Use mapping vmap to find a replacement for v (leave as v if none found) I.e relabelV turns a Mapping into a total function using identity for undefined cases

all3 :: (a -> b) -> (a, a, a) -> (b, b, b)

Applies a function to all three elements of a triple

relabelNewExcept :: [Vertex] -> [Vertex] -> Tgraph -> Tgraph

relabelNewExcept fixed avoid g produces a new Tgraph from g by relabeling. Any vertex in g that is in the list fixed will remain unchanged. All other vertices in g that are in the list avoid will be changed to new vertices that are neither in g nor in the list avoid. (If a vertex in g appears in both fixed and avoid, it will remain unchanged)

newVMapAvoid :: [Vertex] -> [Vertex] -> Mapping Vertex Vertex

newVMapAvoid avoid vs produces a 1-1 mapping from vertices in vs to new vertices such that no new vertex is in the list avoid or in vs

relabelByCommonEdge :: Tgraph -> DEdge -> Tgraph -> Tgraph

relabelByCommonEdge g1 (x,y) g2 produces a relabeled version of g2 that is consistent with g1 on their overlap. The overlaping region must contain the common directed edge (x,y) without relabeling.

CAVEAT: The overlap must be a SINGLE tile-connected region in g1. (If the overlap contains more than one tile-connected region the result may not be a correct relabeling of g2)

relabelByEdges :: (Tgraph, DEdge) -> (Tgraph, DEdge) -> Tgraph

rrelabelByEdges (g1,e1) (g2,e2) produces a relabeled version of g2 that is consistent with g1 on their overlap. The overlaping region must contain the directed edge e1 in g1 and e2 in g2, which will be identified by the relabeling of g2.

CAVEAT: The overlap must be a SINGLE tile-connected region in g1. (If the overlap contains more than one tile-connected region the result may not be a correct relabeling of g2)

createVMapUsing :: (Tgraph, TileFace) -> (Tgraph, TileFace) -> Mapping Vertex Vertex

createVMapUsing (g1,fc1) (g2,fc2) creates a relabeling map to make g2 consistent with g1 in a region of overlap if this is possible. fc1 and fc2 should be matching face types and have at least 2 vertices the same (without relabeling), with fc1 in g1 and fc2 in g2. g2 must have no vertices in common with g1 at the outset except for the 2 common vertices in fc1 and fc2 The resulting map when applied to fc2 will be identical to fc1. The common overlap needs to be a SINGLE region of tile-connected faces in g1. The common region must contain fc1 in g1 and fc2 in g2 The resulting map gives the required relabeling for g2 This will not produce correct results if there are other overlaps not tile-connected in g1 to the region containing fc1

addToVmap :: Tgraph -> [TileFace] -> [TileFace] -> [TileFace] -> Mapping Vertex Vertex -> Mapping Vertex Vertex

addToVmap is used by createVMapUsing to build a relabeling map. (addToVmap g processing tried awaiting vmap) produces a new vmap to be used for relabeling faces being matched against g. A match occurs if a relabeled face matches the face type and at least 2 corresponding vertices of a face in g. Argument g is the Tgraph being matched against. Argument processing is a list of faces to be matched next and each has an edge in common with at least one previously matched face. Argument tried is a list of faces that have been checked for a matching face in g but none found. They may still be on the boundary of g. Argument awaiting is a list of faces that have not yet been tried for a match and are not tile-connected to any faces already matched. Argument vmap is a vertex to vertex mapping used to store relabeling information during the process.

The idea is that from a single matched starting face we process faces that share an edge with a previously matched face. We prioritorize processing faces that have a match in g (with 2 matching vertices). If a face is tried but has no such match, it is stacked up in the tried list to await a check (with addToVmapBdCheck) for boundary cases after all faces in the overlap have been processed (when processing == []).

addToVmapBdCheck :: Boundary -> [TileFace] -> Mapping Vertex Vertex -> Mapping Vertex Vertex

checkForBoundaryThirdV :: Boundary -> TileFace -> Maybe (Vertex, Vertex)

conflictCheck :: TileFace -> Boundary -> Bool

analyse :: DEdge -> TileFace -> (Int, Int, Vertex)

thirdVertexIn :: Tgraph -> TileFace -> Maybe [(Vertex, Vertex)]

thirdVertexIn g f tries to find a face in g corresponding to face f, where at least 2 vertices have to match. If successful, it returns Just [(v,v')] where v needs to be relabeled to v' for a 3 verrtex match, OR it may return Just [] if there is already a 3 vertex match. A return of Nothing indicates a failure to find a match.

pairing :: DEdge -> Tgraph -> Tgraph -> Maybe (TileFace, TileFace)

pairing e g1 g2 - finds a a face in g1 with directed edge and a face in g2 with directed edge e It returns an empty list if the faces do not match,

hasEdge :: DEdge -> TileFace -> Bool

hasEdge e f returns True if directed edge e is one of the directed edges of face f

initVMap :: TileFace -> TileFace -> Mapping Vertex Vertex

initVMap f1 f2 - creates a relabeling (vertex to vertex mapping) so that if applied to face f2, the vertices will match with face f1 exactly.

matchFace :: TileFace -> [TileFace] -> Maybe TileFace

matchFace f1 fcs - looks for a face in fcs that has a directed edge in common with f1. If such a face is found it checks that it matches (except possibly for one vertex label) Returns Just f if f is the matched face, but an error f does not match. Retruns Nothing if no face in fcs has a directed edge in common with f1

hasEdgeIn :: [DEdge] -> TileFace -> Bool

hasEdgeIn es fc - is True if fc has a directed edge in the list of edges es.

twoVMatch :: TileFace -> TileFace -> Bool

twoVMatch f1 f2 is True if the two tilefaces are the same except for a single vertex label possibly not matching.

sameFaceType :: HalfTile rep1 -> HalfTile rep2 -> Bool

sameFaceType f1 f2 is True if f1 and f2 are of the same form (i.e. both LD or both RD or both LK or both RK)

module Tgraph.Prelude

module Tgraph.Decompose

module Tgraph.Compose

module Tgraph.Force

module Tgraph.Convert