Tgraphs

Description

This is the main module for Tgraphs which collects and exports the various other Tgraph modules and includes a definition of emplace and other experimental combinations.

Documentation

emplace :: Tgraph -> Tgraph

emplace does maximal composing with force and composeG, then applies decomposeG and force repeatedly back to the starting level. It produces the emplacement of influence of the argument graph.

emplacements :: Tgraph -> [Tgraph]

emplacements is best supplied with a maximally composed or near maximally composed graph It produces an infinite list of emplacements of the starting graph and its decompositions.

emplaceChoices :: Tgraph -> [Tgraph]

a version of emplace using makeChoices at the top level. after makeChoices we use emplace to attempt further compositions but with no further choices

makeChoices :: Tgraph -> [Tgraph]

makeChoices is a temporary tool which does not attempt to analyse choices for correctness. It can thus create some choices which will be incorrect. The unknowns returned from classifyDartWings can become largeKiteCentres or largeDartBases. This produces 2^n choices where n is the number of unknowns (excluding lone dart wing tips with valencyD 2).

compForce :: Tgraph -> Tgraph

compForce does a force then composeG but it by-passes the check on the composed graph because the argument was forced

allCompFs :: Tgraph -> [Tgraph]

allCompFs g produces a list of all forced compositions starting from g up to but excluding the empty graph

allComps :: Tgraph -> [Tgraph]

allComps g produces a list of all compositions starting from g up to but excluding the empty graph. This is not safe in general as it can fail by producing a non-connected graph or graph with crossing boundaries.

maxCompose :: Tgraph -> Tgraph

maxCompose and maxFCompose produce a maximal graph.

maxFCompose :: Tgraph -> Tgraph

maxCompose and maxFCompose produce a maximal graph.

removeIncompleteTiles :: Tgraph -> Tgraph

remove halftile faces that do not have their matching half tile

module Tgraph.Prelude

module Tgraph.Decompose

module Tgraph.Compose

module Tgraph.Force