TileFigExamples

Description

Documentation

label :: (Renderable (Text n) b, RealFloat n, Typeable n) => String -> Point V2 n -> QDiagram b V2 n Any

Used for adding text at a point

thePieces :: [Piece]

Figures for the four Pieces

piecesFig :: Diagram B

drawn edges of 4 pieces in a row

piecesFig2 :: QDiagram B V2 Double Any

filled 4 pieces in a row

markedTiles :: Diagram B

figure showing origins and markings on tiles

markedTiles2 :: Diagram B

another figure showing origins and markings on tiles

newPiecesFig :: Diagram B

labelled pieces with join and origin shown

tileIntro :: Diagram B

diagram combining markedTiles2 and newPiecesFig

fourDecomps :: Diagram B

4 decompositions in a column for each piece

fiveInflate :: Diagram B

example of inflate in action (5 chosen inflate steps) show inital and final piece together on left patch, and 5 decomposition of final piece in right patch fiveInflate = hsep 1 $ fmap drawPatch [[ld,lk'], multiDecomp 5 [lk']] where -- two seperate patches

inflatefig :: Located Piece -> Diagram B

example of first 5 alternatives of 4-fold inflations of lp (located piece)

fiveAlternatives :: Diagram B

fdiagram showing first five alternatives of 4-fold inflations of a right dart

suns :: [Patch]

An infinite list of patches of increasingly decomposed sun

sun5 :: Patch

sun5 is a patch of a 5 times decomposed sun, and sun6 is a patch of a 6 times decomposed sun

sun6 :: Patch

sun5 is a patch of a 5 times decomposed sun, and sun6 is a patch of a 6 times decomposed sun

sun6Fig :: Diagram B

Diagram for sun6

sun5Over6Fig :: Diagram B

overlaying sun5 in red atop sun6

experimentFig :: Diagram B

Using experiment (defined in Tilelib) on sun6 clearly illustrates the embedded sun5

twoLevelsFig :: Diagram B

similarly experiment on sun4

dartWingFig :: Diagram B

shows two types of dart wing vertices (largeKiteCentre, largeDartBase)

filledSun6 :: Diagram B

using fillDK'

newFillSun6 :: Diagram B

using fillDK

threeShapesSample :: Diagram B

showing three decomposed shapes with colouring for dart,kite and grout (edges) arranged in an arc

threeColourFilled :: Diagram B

showing three decomposed shapes with colouring for dart,kite and grout (edges) arranged in row

type Swatch = [(Colour Double, Colour Double, Colour Double)]

A swatch is a list of triples of colours which are used to fill sun5s (dart, kite, grout)

showSwatch :: Int -> Swatch -> Diagram B

showSwatch n sw produces a diagram from a swatch sw where n is used to indicate how many samples in a row display at 800

showSample :: (Colour Double, Colour Double, Colour Double) -> Diagram B

a sample is similarly a triple of colours for a single sun6 display at 800

showLargeSample :: (Colour Double, Colour Double, Colour Double) -> Diagram B

a large sample is similarly a triple of colours for a single sun7 display at 1000

sL1 :: Diagram B

sample coloured tile diagrams (using decomposed suns)

sL2 :: Diagram B

sample coloured tile diagrams (using decomposed suns)

sL3 :: Diagram B

sample coloured tile diagrams (using decomposed suns)

news1 :: Diagram B

sample coloured tile diagrams (using decomposed suns)

s7 :: Diagram B

sample coloured tile diagrams (using decomposed suns)

s8 :: Diagram B

sample coloured tile diagrams (using decomposed suns)

s9 :: Diagram B

sample coloured tile diagrams (using decomposed suns)

s1 :: Diagram B

sample coloured tile diagrams (using decomposed suns)

s2 :: Diagram B

sample coloured tile diagrams (using decomposed suns)

s3 :: Diagram B

sample coloured tile diagrams (using decomposed suns)

s4 :: Diagram B

sample coloured tile diagrams (using decomposed suns)

s5 :: Diagram B

sample coloured tile diagrams (using decomposed suns)

s6 :: Diagram B

sample coloured tile diagrams (using decomposed suns)

coasterSwatch :: Swatch

example swatches

darkM :: Swatch

example swatches

bluegreens :: Swatch

example swatches

blues :: Swatch

example swatches

pinks :: Swatch

example swatches

pastel :: Swatch

example swatches

greens :: Swatch

example swatches

reds :: Swatch

example swatches

swatch3 :: Swatch

example swatches

swatch2 :: Swatch

example swatches

swatch1 :: Swatch

example swatches